Crate classicube_sys

use classicube_sys::*;
use std::{os::raw::c_int, ptr};

extern "C" fn init() {
  let owned_string = OwnedString::new("hello from rust!");

  unsafe {
    Chat_Add(owned_string.as_cc_string());
  }
}

#[no_mangle]
pub static Plugin_ApiVersion: c_int = 1;

#[no_mangle]
pub static mut Plugin_Component: IGameComponent = IGameComponent {
  /* Called when the game is being loaded. */
  Init: Some(init),
  /* Called when the component is being freed. (e.g. due to game being closed) */
  Free: None,
  /* Called to reset the component's state. (e.g. reconnecting to server) */
  Reset: None,
  /* Called to update the component's state when the user begins loading a new map. */
  OnNewMap: None,
  /* Called to update the component's state when the user has finished loading a new map. */
  OnNewMapLoaded: None,
  /* Next component in linked list of components. */
  next: ptr::null_mut(),
};

Structs

AABB

AnimatedComp

Bitmap_

ChatCommand

Entity

EntityVTABLE

Event_Block

Event_Chat

Event_Entry

Event_Float

Event_Input

Event_Int

Event_PointerMove

Event_String

Event_Void

IGameComponent

IVec3_

LocationUpdate

Matrix

Model

ModelTex

ModelVertex

OwnedChatCommand

OwnedString

ScheduledTask

Stream

String_

StringsBuffer_

Texture

TextureRec_

Vec3_

Vec4

VertexP3fT2fC4b_

_BlockEventsList

_ChatEventsList

_EntitiesData

_EntityEventsList

_EnvData

_GfxEventsList

_InputEventsList

_NetEventsList

_PointerEventsList

_ServerConnectionData

_TabListData

_TabListEventsList

_TextureEventsList

_UserEventsList

_WindowEventsList

_WorldData

_WorldEventsList

Constants

FogFunc__FOG_EXP

FogFunc__FOG_EXP2

FogFunc__FOG_LINEAR

Input_Names

KeyBind_Defaults

KeyBind__KEYBIND_AUTOROTATE

KeyBind__KEYBIND_AXIS_LINES

KeyBind__KEYBIND_BACK

KeyBind__KEYBIND_BREAK_LIQUIDS

KeyBind__KEYBIND_CHAT

KeyBind__KEYBIND_COUNT

KeyBind__KEYBIND_DELETE_BLOCK

KeyBind__KEYBIND_DROP_BLOCK

KeyBind__KEYBIND_EXT_INPUT

KeyBind__KEYBIND_FLY

KeyBind__KEYBIND_FLY_DOWN

KeyBind__KEYBIND_FLY_UP

KeyBind__KEYBIND_FOG

KeyBind__KEYBIND_FORWARD

KeyBind__KEYBIND_FULLSCREEN

KeyBind__KEYBIND_HALF_SPEED

KeyBind__KEYBIND_HIDE_FPS

KeyBind__KEYBIND_HIDE_GUI

KeyBind__KEYBIND_HOTBAR_SWITCH

KeyBind__KEYBIND_IDOVERLAY

KeyBind__KEYBIND_INVENTORY

KeyBind__KEYBIND_JUMP

KeyBind__KEYBIND_LEFT

KeyBind__KEYBIND_NOCLIP

KeyBind__KEYBIND_PICK_BLOCK

KeyBind__KEYBIND_PLACE_BLOCK

KeyBind__KEYBIND_PLAYER_LIST

KeyBind__KEYBIND_RESPAWN

KeyBind__KEYBIND_RIGHT

KeyBind__KEYBIND_SCREENSHOT

KeyBind__KEYBIND_SEND_CHAT

KeyBind__KEYBIND_SET_SPAWN

KeyBind__KEYBIND_SMOOTH_CAMERA

KeyBind__KEYBIND_SPEED

KeyBind__KEYBIND_THIRD_PERSON

KeyBind__KEYBIND_ZOOM_SCROLL

Key_Ascii_Names

Key_Function_Names

Key_INPUT_COUNT

Key_KEY_0

Key_KEY_1

Key_KEY_2

Key_KEY_3

Key_KEY_4

Key_KEY_5

Key_KEY_6

Key_KEY_7

Key_KEY_8

Key_KEY_9

Key_KEY_A

Key_KEY_B

Key_KEY_BACKSLASH

Key_KEY_BACKSPACE

Key_KEY_C

Key_KEY_CAPSLOCK

Key_KEY_COMMA

Key_KEY_D

Key_KEY_DELETE

Key_KEY_DOWN

Key_KEY_E

Key_KEY_END

Key_KEY_ENTER

Key_KEY_EQUALS

Key_KEY_ESCAPE

Key_KEY_F

Key_KEY_F1

Key_KEY_F2

Key_KEY_F3

Key_KEY_F4

Key_KEY_F5

Key_KEY_F6

Key_KEY_F7

Key_KEY_F8

Key_KEY_F9

Key_KEY_F10

Key_KEY_F11

Key_KEY_F12

Key_KEY_F13

Key_KEY_F14

Key_KEY_F15

Key_KEY_F16

Key_KEY_F17

Key_KEY_F18

Key_KEY_F19

Key_KEY_F20

Key_KEY_F21

Key_KEY_F22

Key_KEY_F23

Key_KEY_F24

Key_KEY_F25

Key_KEY_F26

Key_KEY_F27

Key_KEY_F28

Key_KEY_F29

Key_KEY_F30

Key_KEY_F31

Key_KEY_F32

Key_KEY_F33

Key_KEY_F34

Key_KEY_F35

Key_KEY_G

Key_KEY_H

Key_KEY_HOME

Key_KEY_I

Key_KEY_INSERT

Key_KEY_J

Key_KEY_K

Key_KEY_KP0

Key_KEY_KP1

Key_KEY_KP2

Key_KEY_KP3

Key_KEY_KP4

Key_KEY_KP5

Key_KEY_KP6

Key_KEY_KP7

Key_KEY_KP8

Key_KEY_KP9

Key_KEY_KP_DECIMAL

Key_KEY_KP_DIVIDE

Key_KEY_KP_ENTER

Key_KEY_KP_MINUS

Key_KEY_KP_MULTIPLY

Key_KEY_KP_PLUS

Key_KEY_L

Key_KEY_LALT

Key_KEY_LBRACKET

Key_KEY_LCTRL

Key_KEY_LEFT

Key_KEY_LMOUSE

Key_KEY_LSHIFT

Key_KEY_LWIN

Key_KEY_M

Key_KEY_MENU

Key_KEY_MINUS

Key_KEY_MMOUSE

Key_KEY_N

Key_KEY_NONE

Key_KEY_NUMLOCK

Key_KEY_O

Key_KEY_P

Key_KEY_PAGEDOWN

Key_KEY_PAGEUP

Key_KEY_PAUSE

Key_KEY_PERIOD

Key_KEY_PRINTSCREEN

Key_KEY_Q

Key_KEY_QUOTE

Key_KEY_R

Key_KEY_RALT

Key_KEY_RBRACKET

Key_KEY_RCTRL

Key_KEY_RIGHT

Key_KEY_RMOUSE

Key_KEY_RSHIFT

Key_KEY_RWIN

Key_KEY_S

Key_KEY_SCROLLLOCK

Key_KEY_SEMICOLON

Key_KEY_SLASH

Key_KEY_SPACE

Key_KEY_T

Key_KEY_TAB

Key_KEY_TILDE

Key_KEY_U

Key_KEY_UP

Key_KEY_V

Key_KEY_W

Key_KEY_X

Key_KEY_XBUTTON1

Key_KEY_XBUTTON2

Key_KEY_Y

Key_KEY_Z

MatrixType__MATRIX_PROJECTION

MatrixType__MATRIX_TEXTURE

MatrixType__MATRIX_VIEW

MsgType_MSG_TYPE_ANNOUNCEMENT

MsgType_MSG_TYPE_BOTTOMRIGHT_1

MsgType_MSG_TYPE_BOTTOMRIGHT_2

MsgType_MSG_TYPE_BOTTOMRIGHT_3

MsgType_MSG_TYPE_CLIENTSTATUS_1

MsgType_MSG_TYPE_CLIENTSTATUS_2

MsgType_MSG_TYPE_NORMAL

MsgType_MSG_TYPE_STATUS_1

MsgType_MSG_TYPE_STATUS_2

MsgType_MSG_TYPE_STATUS_3

PACKEDCOL_A_MASK

PACKEDCOL_A_SHIFT

PACKEDCOL_B_MASK

PACKEDCOL_B_SHIFT

PACKEDCOL_G_MASK

PACKEDCOL_G_SHIFT

PACKEDCOL_RGB_MASK

PACKEDCOL_R_MASK

PACKEDCOL_R_SHIFT

PACKEDCOL_SHADE_X

PACKEDCOL_SHADE_YMIN

PACKEDCOL_SHADE_Z

PACKEDCOL_WHITE

STRING_SIZE

VertexFormat__VERTEX_FORMAT_P3FC4B

VertexFormat__VERTEX_FORMAT_P3FT2FC4B

keybindNames

Statics

BlockEvents

ChatEvents

Entities

EntityEvents

Env

GfxEvents

InputEvents

NetEvents

PointerEvents

Server

TabList

TabListEvents

TextureEvents

UserEvents

WindowEvents

World

WorldEvents

Functions

Chat_Add⚠

Chat_AddOf⚠

Chat_AddRaw⚠

Chat_Send⚠

Commands_Register⚠

Convert_ParseBool⚠

Convert_ParseFloat⚠

Convert_ParseInt⚠

Convert_ParseUInt8⚠

Convert_ParseUInt16⚠

Convert_ParseUInt64⚠

Env_Reset⚠

Env_SetCloudsCol⚠

Env_SetCloudsHeight⚠

Env_SetCloudsSpeed⚠

Env_SetEdgeBlock⚠

Env_SetEdgeHeight⚠

Env_SetExpFog⚠

Env_SetFogCol⚠

Env_SetShadowCol⚠

Env_SetSidesBlock⚠

Env_SetSidesOffset⚠

Env_SetSkyCol⚠

Env_SetSkyboxCol⚠

Env_SetSkyboxHorSpeed⚠

Env_SetSkyboxVerSpeed⚠

Env_SetSunCol⚠

Env_SetWeather⚠

Env_SetWeatherFade⚠

Env_SetWeatherSpeed⚠

Event_RaiseFloat⚠

Event_RaiseInput⚠

Event_RaiseInt⚠

Event_RaiseVoid⚠

Event_Register⚠

Event_RegisterBlock⚠

Event_RegisterChat⚠

Event_RegisterFloat⚠

Event_RegisterInput⚠

Event_RegisterInt⚠

Event_RegisterPointerMove⚠

Event_RegisterString⚠

Event_RegisterVoid⚠

Event_Unregister⚠

Event_UnregisterBlock⚠

Event_UnregisterChat⚠

Event_UnregisterFloat⚠

Event_UnregisterInput⚠

Event_UnregisterInt⚠

Event_UnregisterPointerMove⚠

Event_UnregisterString⚠

Event_UnregisterVoid⚠

Game_ChangeBlock⚠

Game_UpdateBlock⚠

Gfx_BeginFrame⚠

Gfx_BindIb⚠

Gfx_BindTexture⚠

Gfx_BindVb⚠

Gfx_CalcOrthoMatrix⚠

Gfx_CalcPerspectiveMatrix⚠

Gfx_Clear⚠

Gfx_ClearCol⚠

Gfx_CreateAndLockVb⚠

Gfx_CreateDynamicVb⚠

Gfx_CreateIb⚠

Gfx_CreateTexture⚠

Gfx_CreateVb⚠

Gfx_DeleteIb⚠

Gfx_DeleteTexture⚠

Gfx_DeleteVb⚠

Gfx_DisableMipmaps⚠

Gfx_Draw2DFlat⚠

Gfx_Draw2DGradient⚠

Gfx_Draw2DTexture⚠

Gfx_DrawIndexedVb_TrisT2fC4b⚠

Gfx_DrawVb_IndexedTris⚠

Gfx_DrawVb_IndexedTris_Range⚠

Gfx_DrawVb_Lines⚠

Gfx_EnableMipmaps⚠

Gfx_EndFrame⚠

Gfx_Free⚠

Gfx_GetApiInfo⚠

Gfx_GetFog⚠

Gfx_Init⚠

Gfx_LoadIdentityMatrix⚠

Gfx_LoadMatrix⚠

Gfx_LockDynamicVb⚠

Gfx_LockVb⚠

Gfx_LoseContext⚠

Gfx_Make2DQuad⚠

Gfx_MakeIndices⚠

Gfx_Mode2D⚠

Gfx_Mode3D⚠

Gfx_OnWindowResize⚠

Gfx_RestoreAlphaState⚠

Gfx_SetAlphaArgBlend⚠

Gfx_SetAlphaBlending⚠

Gfx_SetAlphaTest⚠

Gfx_SetColWriteMask⚠

Gfx_SetDepthTest⚠

Gfx_SetDepthWrite⚠

Gfx_SetDynamicVbData⚠

Gfx_SetFaceCulling⚠

Gfx_SetFog⚠

Gfx_SetFogCol⚠

Gfx_SetFogDensity⚠

Gfx_SetFogEnd⚠

Gfx_SetFogMode⚠

Gfx_SetFpsLimit⚠

Gfx_SetTexturing⚠

Gfx_SetVertexFormat⚠

Gfx_SetupAlphaState⚠

Gfx_TakeScreenshot⚠

Gfx_TryRestoreContext⚠

Gfx_UnlockDynamicVb⚠

Gfx_UnlockVb⚠

Gfx_UpdateDynamicVb_IndexedTris⚠

Gfx_UpdateTexturePart⚠

Gfx_WarnIfNecessary⚠

Options_Get⚠

Options_GetBool⚠

Options_GetEnum⚠

Options_GetFloat⚠

Options_GetInt⚠

Options_Load⚠

Options_Save⚠

Options_Set⚠

Options_SetBool⚠

Options_SetInt⚠

Options_SetString⚠

PackedCol_A_Bits

PackedCol_B_Bits

PackedCol_G_Bits

PackedCol_Lerp⚠

PackedCol_Make

PackedCol_R_Bits

PackedCol_Scale⚠

PackedCol_Tint⚠

String_Append⚠

String_AppendBool⚠

String_AppendColorless⚠

String_AppendConst⚠

String_AppendFloat⚠

String_AppendHex⚠

String_AppendInt⚠

String_AppendPaddedInt⚠

String_AppendString⚠

String_AppendUInt32⚠

String_CalcLen⚠

String_CaselessContains⚠

String_CaselessEnds⚠

String_CaselessEquals⚠

String_CaselessEqualsConst⚠

String_CaselessStarts⚠

String_Compare⚠

String_Copy⚠

String_CopyToRaw⚠

String_DeleteAt⚠

String_Equals⚠

String_Format1⚠

String_Format2⚠

String_Format3⚠

String_Format4⚠

String_FromReadonly⚠

String_IndexOfAt⚠

String_Init⚠

String_InsertAt⚠

String_LastIndexOfAt⚠

String_MakeUInt32⚠

String_StripCols⚠

String_UNSAFE_Separate⚠

String_UNSAFE_Split⚠

String_UNSAFE_SplitBy⚠

String_UNSAFE_Substring⚠

String_UNSAFE_SubstringAt⚠

String_UNSAFE_TrimEnd⚠

String_UNSAFE_TrimStart⚠

StringsBuffer_Add⚠

StringsBuffer_Clear⚠

StringsBuffer_Remove⚠

StringsBuffer_UNSAFE_Get⚠

TabList_Remove⚠

TabList_Set⚠

World_ApplyTexturePack⚠

World_Blocks

World_Blocks2

World_GetBlock

World_Pack

World_Reset⚠

World_SetNewMap⚠

Type Definitions

Bitmap

BlockID

BlockRaw

EntityID

Event_Block_Callback

Event_Chat_Callback

Event_Entry_Callback

Event_Float_Callback

Event_Input_Callback

Event_Int_Callback

Event_PointerMove_Callback

Event_String_Callback

Event_Void_Callback

FogFunc

FogFunc_

GfxResourceID

IVec3

Key

KeyBind

KeyBind_

MatrixType

MatrixType_

MsgType

PackedCol

String

StringsBuffer

TextureRec

Vec3

VertexFormat

VertexFormat_

VertexP3fT2fC4b

cc_bool

cc_int8

cc_int16

cc_int32

cc_int64

cc_result

cc_uint8

cc_uint16

cc_uint32

cc_uint64

cc_uintptr