Struct classicube_sys::Matrix

#[repr(C)]
pub struct Matrix {
    pub row1: Vec4,
    pub row2: Vec4,
    pub row3: Vec4,
    pub row4: Vec4,
}

Fields
row1: Vec4row2: Vec4row3: Vec4row4: Vec4
Implementations

impl Matrix

pub const IDENTITY: Self = _

pub const fn identity_value() -> Self

pub fn rotate_x(angle: c_float) -> Self

Returns a matrix representing a counter-clockwise rotation around X axis.

pub fn rotate_y(angle: c_float) -> Self

Returns a matrix representing a counter-clockwise rotation around Y axis.

pub fn rotate_z(angle: c_float) -> Self

Returns a matrix representing a counter-clockwise rotation around Z axis.

pub fn translate(x: c_float, y: c_float, z: c_float) -> Self

Returns a matrix representing a translation to the given coordinates.

pub fn scale(x: c_float, y: c_float, z: c_float) -> Self

Returns a matrix representing a scaling by the given factors.

pub fn orthographic( left: c_float, right: c_float, top: c_float, bottom: c_float, zNear: c_float, zFar: c_float ) -> Self

pub fn perspective_field_of_view( fovy: c_float, aspect: c_float, z_near: c_float, z_far: c_float ) -> Self

pub fn look_rot(pos: Vec3, rot: Vec2) -> Self