Type Definition classicube_sys::Vec3

type Vec3 = Vec3_;

Implementations

impl Vec3

pub const fn new(x: c_float, y: c_float, z: c_float) -> Self

pub const fn big_pos() -> Self

pub fn get_dir_vector(yaw_rad: c_float, pitch_rad: c_float) -> Self

pub fn set(&mut self, x: c_float, y: c_float, z: c_float)

Sets the X, Y, and Z components of a 3D vector

pub fn is_zero(&self) -> bool

Whether all components of a 3D vector are 0

pub fn length_squared(&self) -> c_float

Returns the squared length of the vector.

Squared length can be used for comparison, to avoid a costly sqrt()

However, you must sqrt() this when adding lengths.

pub const fn floor(&self) -> IVec3

pub fn negate(&self) -> Self

Negates the components of a vector.