Crate classicube_sys

use classicube_sys::*;
use std::{os::raw::c_int, ptr};

extern "C" fn init() {
  let owned_string = OwnedString::new("hello from rust!");

  unsafe {
    Chat_Add(owned_string.as_cc_string());
  }
}

#[no_mangle]
pub static Plugin_ApiVersion: c_int = 1;

#[no_mangle]
pub static mut Plugin_Component: IGameComponent = IGameComponent {
  /* Called when the game is being loaded. */
  Init: Some(init),
  /* Called when the component is being freed. (e.g. due to game being closed) */
  Free: None,
  /* Called to reset the component's state. (e.g. reconnecting to server) */
  Reset: None,
  /* Called to update the component's state when the user begins loading a new map. */
  OnNewMap: None,
  /* Called to update the component's state when the user has finished loading a new map. */
  OnNewMapLoaded: None,
  /* Next component in linked list of components. */
  next: ptr::null_mut(),
};

Structs

AABB

AnimatedComp

ChatCommand

Entity

EntityVTABLE

Event_Block

Event_Chat

Event_Entry

Event_Float

Event_Input

Event_Int

Event_PointerMove

Event_Void

IGameComponent

IVec3_

LocationUpdate

Matrix

Model

OwnedChatCommand

OwnedString

PackedCol___bindgen_ty_1

ScheduledTask

Stream

String_

StringsBuffer_

Texture

TextureRec_

Vec3_

Vec4

_BlockEventsList

_ChatEventsList

_EntitiesData

_EntityEventsList

_GfxEventsList

_KeyEventsList

_NetEventsList

_PointerEventsList

_ServerConnectionData

_TabListData

_TabListEventsList

_TextureEventsList

_UserEventsList

_WindowEventsList

_WorldEventsList

Constants

Key__INPUT_COUNT

Key__KEY_0

Key__KEY_1

Key__KEY_2

Key__KEY_3

Key__KEY_4

Key__KEY_5

Key__KEY_6

Key__KEY_7

Key__KEY_8

Key__KEY_9

Key__KEY_A

Key__KEY_B

Key__KEY_BACKSLASH

Key__KEY_BACKSPACE

Key__KEY_C

Key__KEY_CAPSLOCK

Key__KEY_COMMA

Key__KEY_D

Key__KEY_DELETE

Key__KEY_DOWN

Key__KEY_E

Key__KEY_END

Key__KEY_ENTER

Key__KEY_EQUALS

Key__KEY_ESCAPE

Key__KEY_F

Key__KEY_F1

Key__KEY_F2

Key__KEY_F3

Key__KEY_F4

Key__KEY_F5

Key__KEY_F6

Key__KEY_F7

Key__KEY_F8

Key__KEY_F9

Key__KEY_F10

Key__KEY_F11

Key__KEY_F12

Key__KEY_F13

Key__KEY_F14

Key__KEY_F15

Key__KEY_F16

Key__KEY_F17

Key__KEY_F18

Key__KEY_F19

Key__KEY_F20

Key__KEY_F21

Key__KEY_F22

Key__KEY_F23

Key__KEY_F24

Key__KEY_F25

Key__KEY_F26

Key__KEY_F27

Key__KEY_F28

Key__KEY_F29

Key__KEY_F30

Key__KEY_F31

Key__KEY_F32

Key__KEY_F33

Key__KEY_F34

Key__KEY_F35

Key__KEY_G

Key__KEY_H

Key__KEY_HOME

Key__KEY_I

Key__KEY_INSERT

Key__KEY_J

Key__KEY_K

Key__KEY_KP0

Key__KEY_KP1

Key__KEY_KP2

Key__KEY_KP3

Key__KEY_KP4

Key__KEY_KP5

Key__KEY_KP6

Key__KEY_KP7

Key__KEY_KP8

Key__KEY_KP9

Key__KEY_KP_DECIMAL

Key__KEY_KP_DIVIDE

Key__KEY_KP_ENTER

Key__KEY_KP_MINUS

Key__KEY_KP_MULTIPLY

Key__KEY_KP_PLUS

Key__KEY_L

Key__KEY_LALT

Key__KEY_LBRACKET

Key__KEY_LCTRL

Key__KEY_LEFT

Key__KEY_LMOUSE

Key__KEY_LSHIFT

Key__KEY_LWIN

Key__KEY_M

Key__KEY_MENU

Key__KEY_MINUS

Key__KEY_MMOUSE

Key__KEY_N

Key__KEY_NONE

Key__KEY_NUMLOCK

Key__KEY_O

Key__KEY_P

Key__KEY_PAGEDOWN

Key__KEY_PAGEUP

Key__KEY_PAUSE

Key__KEY_PERIOD

Key__KEY_PRINTSCREEN

Key__KEY_Q

Key__KEY_QUOTE

Key__KEY_R

Key__KEY_RALT

Key__KEY_RBRACKET

Key__KEY_RCTRL

Key__KEY_RIGHT

Key__KEY_RMOUSE

Key__KEY_RSHIFT

Key__KEY_RWIN

Key__KEY_S

Key__KEY_SCROLLLOCK

Key__KEY_SEMICOLON

Key__KEY_SLASH

Key__KEY_SPACE

Key__KEY_T

Key__KEY_TAB

Key__KEY_TILDE

Key__KEY_U

Key__KEY_UP

Key__KEY_V

Key__KEY_W

Key__KEY_X

Key__KEY_XBUTTON1

Key__KEY_XBUTTON2

Key__KEY_Y

Key__KEY_Z

MsgType_MSG_TYPE_ANNOUNCEMENT

MsgType_MSG_TYPE_BOTTOMRIGHT_1

MsgType_MSG_TYPE_BOTTOMRIGHT_2

MsgType_MSG_TYPE_BOTTOMRIGHT_3

MsgType_MSG_TYPE_CLIENTSTATUS_1

MsgType_MSG_TYPE_CLIENTSTATUS_2

MsgType_MSG_TYPE_NORMAL

MsgType_MSG_TYPE_STATUS_1

MsgType_MSG_TYPE_STATUS_2

MsgType_MSG_TYPE_STATUS_3

STRING_SIZE

Statics

BlockEvents

ChatEvents

Entities

EntityEvents

GfxEvents

InputEvents

NetEvents

PointerEvents

Server

TabList

TabListEvents

TextureEvents

UserEvents

WindowEvents

WorldEvents

Functions

Chat_Add⚠

Chat_AddOf⚠

Chat_AddRaw⚠

Chat_Send⚠

Commands_Register⚠

Convert_ParseBool⚠

Convert_ParseFloat⚠

Convert_ParseInt⚠

Convert_ParseUInt8⚠

Convert_ParseUInt16⚠

Convert_ParseUInt64⚠

Event_RaiseFloat⚠

Event_RaiseInput⚠

Event_RaiseInt⚠

Event_RaiseVoid⚠

Event_Register⚠

Event_RegisterChat⚠

Event_RegisterFloat⚠

Event_RegisterInput⚠

Event_RegisterInt⚠

Event_RegisterVoid⚠

Event_Unregister⚠

Event_UnregisterChat⚠

Event_UnregisterFloat⚠

Event_UnregisterInput⚠

Event_UnregisterInt⚠

Event_UnregisterVoid⚠

Options_Get⚠

Options_GetBool⚠

Options_GetEnum⚠

Options_GetFloat⚠

Options_GetInt⚠

Options_Load⚠

Options_Save⚠

Options_Set⚠

Options_SetBool⚠

Options_SetInt⚠

Options_SetString⚠

String_Append⚠

String_AppendBool⚠

String_AppendColorless⚠

String_AppendConst⚠

String_AppendFloat⚠

String_AppendHex⚠

String_AppendInt⚠

String_AppendPaddedInt⚠

String_AppendString⚠

String_AppendUInt32⚠

String_CalcLen⚠

String_CaselessContains⚠

String_CaselessEnds⚠

String_CaselessEquals⚠

String_CaselessEqualsConst⚠

String_CaselessStarts⚠

String_Compare⚠

String_Copy⚠

String_CopyToRaw⚠

String_DeleteAt⚠

String_Equals⚠

String_Format1⚠

String_Format2⚠

String_Format3⚠

String_Format4⚠

String_FromReadonly⚠

String_IndexOfAt⚠

String_IndexOfString⚠

String_Init⚠

String_InsertAt⚠

String_LastIndexOfAt⚠

String_MakeUInt32⚠

String_StripCols⚠

String_UNSAFE_Separate⚠

String_UNSAFE_Split⚠

String_UNSAFE_SplitBy⚠

String_UNSAFE_Substring⚠

String_UNSAFE_SubstringAt⚠

String_UNSAFE_TrimEnd⚠

String_UNSAFE_TrimStart⚠

StringsBuffer_Add⚠

StringsBuffer_Clear⚠

StringsBuffer_Remove⚠

StringsBuffer_UNSAFE_Get⚠

TabList_Remove⚠

TabList_Set⚠

Type Definitions

BlockID

EntityID

Event_Block_Callback

Event_Chat_Callback

Event_Entry_Callback

Event_Float_Callback

Event_Input_Callback

Event_Int_Callback

Event_PointerMove_Callback

Event_Void_Callback

GfxResourceID

IVec3

Key

Key_

MsgType

PackedCol

String

StringsBuffer

TextureRec

Vec3

bool_

cc_int16

cc_uint8

cc_uint16

cc_uint32

cc_uint64

cc_uintptr

Unions

PackedCol_