Crate classicube_sys

use classicube_sys::*;
use std::{os::raw::c_int, ptr};

extern "C" fn init() {
  let owned_string = OwnedString::new("hello from rust!");

  unsafe {
    Chat_Add(owned_string.as_cc_string());
  }
}

#[no_mangle]
pub static Plugin_ApiVersion: c_int = 1;

#[no_mangle]
pub static mut Plugin_Component: IGameComponent = IGameComponent {
  /* Called when the game is being loaded. */
  Init: Some(init),
  /* Called when the component is being freed. (e.g. due to game being closed) */
  Free: None,
  /* Called to reset the component's state. (e.g. reconnecting to server) */
  Reset: None,
  /* Called to update the component's state when the user begins loading a new map. */
  OnNewMap: None,
  /* Called to update the component's state when the user has finished loading a new map. */
  OnNewMapLoaded: None,
  /* Next component in linked list of components. */
  next: ptr::null_mut(),
};

Structs

AABB

AnimatedComp

AudioFormat

Bitmap_

BoxDesc

ButtonWidget

Camera

ChatCommand

ChatInputWidget

ChatScreen

CheckUpdateData

ChunkInfo

ChunkPartInfo

Codebook

CollisionsComp

CustomParticleEffect

DateTime

DeflateState

DisplayDevice

DrawTextArgs

Entity

EntityVTABLE

EntryList

Event_Block

Event_Chat

Event_Entry

Event_Float

Event_Input

Event_Int

Event_PointerMove

Event_String

Event_Void

FetchFlagsData

FetchServerData

FetchServersData

FetchUpdateData

Floor

FontDesc

GZipHeader

GZipState

GetTokenTaskData

GraphicsMode

HUDScreen

HacksComp

HotbarWidget

HotkeyData

HttpRequest

HuffmanTable

IGameComponent

IVec3_

InflateState

InputWidget

InterpComp

InterpState

JsonContext

LButton

LInput

LLabel

LLine

LScreen

LSlider

LTable

LTableColumn

LWebTask

LWidget

LWidgetVTABLE

LocalPlayer

LocationUpdate

Mapping

Matrix

MenuInputDesc

MenuInputDesc__bindgen_ty_1__bindgen_ty_1

MenuInputDesc__bindgen_ty_1__bindgen_ty_2

MenuInputDesc__bindgen_ty_1__bindgen_ty_3

MenuInputDesc__bindgen_ty_1__bindgen_ty_4

MenuInputVTABLE

MenuInputWidget

Mode

Model

ModelPart

ModelTex

ModelVertex

NetInterpComp

NetPlayer

OggState

OwnedChatCommand

OwnedString

Particle

PhysicsComp

Physics_

PlayerListWidget

Pointer

RayTracer

Rect2D_

Residue

ScheduledTask

Screen

ScreenVTABLE

ScrollbarWidget

SearcherState

ServerInfo

SignInTaskData

Size2D_

SpecialInputTab

SpecialInputWidget

Stream

Stream__bindgen_ty_1__bindgen_ty_1

Stream__bindgen_ty_1__bindgen_ty_2

Stream__bindgen_ty_1__bindgen_ty_3

Stream__bindgen_ty_1__bindgen_ty_4

Stream__bindgen_ty_1__bindgen_ty_5

String_

StringsBuffer_

TableWidget

TextAtlas

TextGroupWidget

TextWidget

Texture

TextureRec_

TiltComp

Vec2_

Vec3_

Vec4

VertexP3fC4b_

VertexP3fT2fC4b_

VorbisState

VorbisWindow

Widget

WidgetVTABLE

ZLibHeader

ZLibState

ZipEntry

ZipState

_Atlas1DData

_Atlas2DData

_BlockEventsList

_BlockLists

_CameraData

_ChatEventsList

_DrawerData

_EntitiesData

_EntityEventsList

_EnvData

_GameData

_GfxData

_GfxEventsList

_InputEventsList

_InventoryData

_ModelsData

_NetEventsList

_PointerEventsList

_ServerConnectionData

_TabListData

_TabListEventsList

_TextureEventsList

_UserEventsList

_WindowEventsList

_WorldData

_WorldEventsList

__BindgenBitfieldUnit

imdct_state

Constants

ATLAS1D_MAX_ATLASES

ATLAS2D_TILES_PER_ROW

ATLAS2D_MASK

ATLAS2D_SHIFT

ATLAS2D_MAX_ROWS_COUNT

AUDIO_MAX_BUFFERS

BITMAPCOL_A_MASK

BITMAPCOL_A_SHIFT

BITMAPCOL_B_MASK

BITMAPCOL_B_SHIFT

BITMAPCOL_G_MASK

BITMAPCOL_G_SHIFT

BITMAPCOL_RGB_MASK

BITMAPCOL_R_MASK

BITMAPCOL_R_SHIFT

BLOCKID_BLOCK_AIR

BLOCKID_BLOCK_AQUA

BLOCKID_BLOCK_BEDROCK

BLOCKID_BLOCK_BLACK

BLOCKID_BLOCK_BLUE

BLOCKID_BLOCK_BOOKSHELF

BLOCKID_BLOCK_BRICK

BLOCKID_BLOCK_BROWN

BLOCKID_BLOCK_BROWN_SHROOM

BLOCKID_BLOCK_CERAMIC_TILE

BLOCKID_BLOCK_COAL_ORE

BLOCKID_BLOCK_COBBLE

BLOCKID_BLOCK_COBBLE_SLAB

BLOCKID_BLOCK_COUNT

BLOCKID_BLOCK_CPE_COUNT

BLOCKID_BLOCK_CRATE

BLOCKID_BLOCK_CYAN

BLOCKID_BLOCK_DANDELION

BLOCKID_BLOCK_DEEP_BLUE

BLOCKID_BLOCK_DIRT

BLOCKID_BLOCK_DOUBLE_SLAB

BLOCKID_BLOCK_FIRE

BLOCKID_BLOCK_FOREST_GREEN

BLOCKID_BLOCK_GLASS

BLOCKID_BLOCK_GOLD

BLOCKID_BLOCK_GOLD_ORE

BLOCKID_BLOCK_GRASS

BLOCKID_BLOCK_GRAVEL

BLOCKID_BLOCK_GRAY

BLOCKID_BLOCK_GREEN

BLOCKID_BLOCK_ICE

BLOCKID_BLOCK_INDIGO

BLOCKID_BLOCK_IRON

BLOCKID_BLOCK_IRON_ORE

BLOCKID_BLOCK_LAVA

BLOCKID_BLOCK_LEAVES

BLOCKID_BLOCK_LIGHT_PINK

BLOCKID_BLOCK_LIME

BLOCKID_BLOCK_LOG

BLOCKID_BLOCK_MAGENTA

BLOCKID_BLOCK_MAGMA

BLOCKID_BLOCK_MAX_CPE

BLOCKID_BLOCK_MAX_DEFINED

BLOCKID_BLOCK_MAX_ORIGINAL

BLOCKID_BLOCK_MOSSY_ROCKS

BLOCKID_BLOCK_OBSIDIAN

BLOCKID_BLOCK_ORANGE

BLOCKID_BLOCK_ORIGINAL_COUNT

BLOCKID_BLOCK_PILLAR

BLOCKID_BLOCK_PINK

BLOCKID_BLOCK_RED

BLOCKID_BLOCK_RED_SHROOM

BLOCKID_BLOCK_ROPE

BLOCKID_BLOCK_ROSE

BLOCKID_BLOCK_SAND

BLOCKID_BLOCK_SANDSTONE

BLOCKID_BLOCK_SAPLING

BLOCKID_BLOCK_SLAB

BLOCKID_BLOCK_SNOW

BLOCKID_BLOCK_SPONGE

BLOCKID_BLOCK_STILL_LAVA

BLOCKID_BLOCK_STILL_WATER

BLOCKID_BLOCK_STONE

BLOCKID_BLOCK_STONE_BRICK

BLOCKID_BLOCK_TEAL

BLOCKID_BLOCK_TNT

BLOCKID_BLOCK_TURQUOISE

BLOCKID_BLOCK_VIOLET

BLOCKID_BLOCK_WATER

BLOCKID_BLOCK_WHITE

BLOCKID_BLOCK_WOOD

BLOCKID_BLOCK_YELLOW

BUTTONWIDGET_MAX

CHUNK_MASK

CHUNK_MAX

CHUNK_SHIFT

CHUNK_SIZE

CHUNK_SIZE_2

CHUNK_SIZE_3

CollideType_COLLIDE_CLIMB_ROPE

CollideType_COLLIDE_GAS

CollideType_COLLIDE_ICE

CollideType_COLLIDE_LIQUID

CollideType_COLLIDE_LIQUID_LAVA

CollideType_COLLIDE_LIQUID_WATER

CollideType_COLLIDE_SLIPPERY_ICE

CollideType_COLLIDE_SOLID

DEFLATE_BLOCK_SIZE

DEFLATE_BUFFER_SIZE

DEFLATE_HASH_MASK

DEFLATE_HASH_SIZE

DEFLATE_OUT_SIZE

DRAWER2D_MAX_COLS

DrawType_DRAW_GAS

DrawType_DRAW_OPAQUE

DrawType_DRAW_SPRITE

DrawType_DRAW_TRANSLUCENT

DrawType_DRAW_TRANSPARENT

DrawType_DRAW_TRANSPARENT_THICK

ENTITIES_MAX_COUNT

ENTITIES_SELF_ID

ENTITY_ADJUSTMENT

ENV_LEGACY

ENV_MINIMAL

ERRORS_ALL_CW_ERR_ROOT_TAG

ERRORS_ALL_CW_ERR_STRING_LEN

ERRORS_ALL_DAT_ERR_IDENTIFIER

ERRORS_ALL_DAT_ERR_JARRAY_CONTENT

ERRORS_ALL_DAT_ERR_JARRAY_TYPE

ERRORS_ALL_DAT_ERR_JCLASS_ANNOTATION

ERRORS_ALL_DAT_ERR_JCLASS_FIELDS

ERRORS_ALL_DAT_ERR_JCLASS_TYPE

ERRORS_ALL_DAT_ERR_JFIELD_CLASS_NAME

ERRORS_ALL_DAT_ERR_JIDENTIFIER

ERRORS_ALL_DAT_ERR_JOBJECT_TYPE

ERRORS_ALL_DAT_ERR_JSTRING_LEN

ERRORS_ALL_DAT_ERR_JVERSION

ERRORS_ALL_DAT_ERR_ROOT_TYPE

ERRORS_ALL_DAT_ERR_VERSION

ERRORS_ALL_ERROR_BASE

ERRORS_ALL_ERR_END_OF_STREAM

ERRORS_ALL_ERR_INVALID_ARGUMENT

ERRORS_ALL_ERR_NOT_SUPPORTED

ERRORS_ALL_ERR_OUT_OF_MEMORY

ERRORS_ALL_FCM_ERR_IDENTIFIER

ERRORS_ALL_FCM_ERR_REVISION

ERRORS_ALL_GZIP_ERR_FLAGS

ERRORS_ALL_GZIP_ERR_HEADER1

ERRORS_ALL_GZIP_ERR_HEADER2

ERRORS_ALL_GZIP_ERR_METHOD

ERRORS_ALL_LVL_ERR_VERSION

ERRORS_ALL_NBT_ERR_INT32S

ERRORS_ALL_NBT_ERR_UNKNOWN

ERRORS_ALL_OGG_ERR_INVALID_SIG

ERRORS_ALL_OGG_ERR_VERSION

ERRORS_ALL_PNG_ERR_COMP_METHOD

ERRORS_ALL_PNG_ERR_FILTER

ERRORS_ALL_PNG_ERR_INTERLACED

ERRORS_ALL_PNG_ERR_INVALID_COL_BPP

ERRORS_ALL_PNG_ERR_INVALID_END_SIZE

ERRORS_ALL_PNG_ERR_INVALID_HDR_SIZE

ERRORS_ALL_PNG_ERR_INVALID_SCANLINE

ERRORS_ALL_PNG_ERR_INVALID_SIG

ERRORS_ALL_PNG_ERR_NO_DATA

ERRORS_ALL_PNG_ERR_PAL_SIZE

ERRORS_ALL_PNG_ERR_TOO_TALL

ERRORS_ALL_PNG_ERR_TOO_WIDE

ERRORS_ALL_PNG_ERR_TRANS_COUNT

ERRORS_ALL_PNG_ERR_TRANS_INVALID

ERRORS_ALL_VORBIS_ERR_BLOCKSIZE

ERRORS_ALL_VORBIS_ERR_CHANS

ERRORS_ALL_VORBIS_ERR_CODEBOOK_ENTRY

ERRORS_ALL_VORBIS_ERR_CODEBOOK_LOOKUP

ERRORS_ALL_VORBIS_ERR_CODEBOOK_SYNC

ERRORS_ALL_VORBIS_ERR_FLOOR_TYPE

ERRORS_ALL_VORBIS_ERR_FRAME_TYPE

ERRORS_ALL_VORBIS_ERR_FRAMING

ERRORS_ALL_VORBIS_ERR_HEADER

ERRORS_ALL_VORBIS_ERR_MAPPING_CHANS

ERRORS_ALL_VORBIS_ERR_MAPPING_RESERVED

ERRORS_ALL_VORBIS_ERR_MAPPING_TYPE

ERRORS_ALL_VORBIS_ERR_MODE_TRANSFORM

ERRORS_ALL_VORBIS_ERR_MODE_TYPE

ERRORS_ALL_VORBIS_ERR_MODE_WINDOW

ERRORS_ALL_VORBIS_ERR_RESIDUE_TYPE

ERRORS_ALL_VORBIS_ERR_TIME_TYPE

ERRORS_ALL_VORBIS_ERR_VERSION

ERRORS_ALL_VORBIS_ERR_WRONG_HEADER

ERRORS_ALL_WAV_ERR_DATA_TYPE

ERRORS_ALL_WAV_ERR_NO_DATA

ERRORS_ALL_WAV_ERR_SAMPLE_BITS

ERRORS_ALL_WAV_ERR_STREAM_HDR

ERRORS_ALL_WAV_ERR_STREAM_TYPE

ERRORS_ALL_ZIP_ERR_FILENAME_LEN

ERRORS_ALL_ZIP_ERR_INVALID_CENTRAL_DIR

ERRORS_ALL_ZIP_ERR_INVALID_LOCAL_DIR

ERRORS_ALL_ZIP_ERR_NO_END_OF_CENTRAL_DIR

ERRORS_ALL_ZIP_ERR_SEEK_CENTRAL_DIR

ERRORS_ALL_ZIP_ERR_SEEK_END_OF_CENTRAL_DIR

ERRORS_ALL_ZIP_ERR_SEEK_LOCAL_DIR

ERRORS_ALL_ZIP_ERR_TOO_MANY_ENTRIES

ERRORS_ALL_ZLIB_ERR_FLAGS

ERRORS_ALL_ZLIB_ERR_METHOD

ERRORS_ALL_ZLIB_ERR_WINDOW_SIZE

EVENT_MAX_CALLBACKS

EXTCHUNK_SIZE

EXTCHUNK_SIZE_2

EXTCHUNK_SIZE_3

EntityType_ENTITY_TYPE_NONE

EntityType_ENTITY_TYPE_PLAYER

EnvVar_ENV_VAR_CLOUDS_COL

EnvVar_ENV_VAR_CLOUDS_HEIGHT

EnvVar_ENV_VAR_CLOUDS_SPEED

EnvVar_ENV_VAR_EDGE_BLOCK

EnvVar_ENV_VAR_EDGE_HEIGHT

EnvVar_ENV_VAR_EXP_FOG

EnvVar_ENV_VAR_FOG_COL

EnvVar_ENV_VAR_SHADOW_COL

EnvVar_ENV_VAR_SIDES_BLOCK

EnvVar_ENV_VAR_SIDES_OFFSET

EnvVar_ENV_VAR_SKYBOX_COL

EnvVar_ENV_VAR_SKYBOX_HOR_SPEED

EnvVar_ENV_VAR_SKYBOX_VER_SPEED

EnvVar_ENV_VAR_SKY_COL

EnvVar_ENV_VAR_SUN_COL

EnvVar_ENV_VAR_WEATHER

EnvVar_ENV_VAR_WEATHER_FADE

EnvVar_ENV_VAR_WEATHER_SPEED

FACE_CONSTS_FACE_COUNT

FACE_CONSTS_FACE_XMAX

FACE_CONSTS_FACE_XMIN

FACE_CONSTS_FACE_YMAX

FACE_CONSTS_FACE_YMIN

FACE_CONSTS_FACE_ZMAX

FACE_CONSTS_FACE_ZMIN

FILENAME_SIZE

FONT_STYLE_FONT_FLAG_UNDERLINE

FONT_STYLE_FONT_STYLE_BOLD

FONT_STYLE_FONT_STYLE_ITALIC

FONT_STYLE_FONT_STYLE_NORMAL

File_SeekFrom_FILE_SEEKFROM_BEGIN

File_SeekFrom_FILE_SEEKFROM_CURRENT

File_SeekFrom_FILE_SEEKFROM_END

FogFunc__FOG_EXP

FogFunc__FOG_EXP2

FogFunc__FOG_LINEAR

FpsLimitMethod_FPS_LIMIT_30

FpsLimitMethod_FPS_LIMIT_60

FpsLimitMethod_FPS_LIMIT_120

FpsLimitMethod_FPS_LIMIT_144

FpsLimitMethod_FPS_LIMIT_COUNT

FpsLimitMethod_FPS_LIMIT_NONE

FpsLimitMethod_FPS_LIMIT_VSYNC

GAME_API_VER

GAME_APP_NAME

GAME_APP_TITLE

GAME_APP_VER

GAME_DEF_TICKS

GAME_MAX_CMDARGS

GAME_NET_TICKS

GFX_APIINFO_LINES

GFX_MAX_INDICES

GFX_MAX_VERTICES

GUI_MAX_SCREENS

GuiAnchor_ANCHOR_CENTRE

GuiAnchor_ANCHOR_CENTRE_MAX

GuiAnchor_ANCHOR_CENTRE_MIN

GuiAnchor_ANCHOR_MAX

GuiAnchor_ANCHOR_MIN

GuiPriority_GUI_PRIORITY_CHAT

GuiPriority_GUI_PRIORITY_DISCONNECT

GuiPriority_GUI_PRIORITY_HUD

GuiPriority_GUI_PRIORITY_INVENTORY

GuiPriority_GUI_PRIORITY_LOADING

GuiPriority_GUI_PRIORITY_MENU

GuiPriority_GUI_PRIORITY_OLDLOADING

GuiPriority_GUI_PRIORITY_TEXIDS

GuiPriority_GUI_PRIORITY_TEXPACK

GuiPriority_GUI_PRIORITY_TOUCH

GuiPriority_GUI_PRIORITY_TOUCHMORE

GuiPriority_GUI_PRIORITY_URLWARNING

HALF_CHUNK_SIZE

HOTKEYS_MAX_COUNT

HotkeyModifiers_HOTKEY_MOD_ALT

HotkeyModifiers_HOTKEY_MOD_CTRL

HotkeyModifiers_HOTKEY_MOD_SHIFT

HttpProgress_ASYNC_PROGRESS_FETCHING_DATA

HttpProgress_ASYNC_PROGRESS_MAKING_REQUEST

HttpProgress_ASYNC_PROGRESS_NOTHING

HttpRequestType_REQUEST_TYPE_GET

HttpRequestType_REQUEST_TYPE_HEAD

HttpRequestType_REQUEST_TYPE_POST

INFLATE_FAST_BITS

INFLATE_MAX_BITS

INFLATE_MAX_CODELENS

INFLATE_MAX_DISTS

INFLATE_MAX_INPUT

INFLATE_MAX_LITS

INFLATE_MAX_LITS_DISTS

INFLATE_WINDOW_MASK

INFLATE_WINDOW_SIZE

INPUTWIDGET_LEN

INPUTWIDGET_MAX_LINES

INPUT_MAX_POINTERS

INVENTORY_BLOCKS_PER_HOTBAR

INVENTORY_HOTBARS

ISOMETRICDRAWER_MAXVERTICES

Input_Names

KeyBind_Defaults

KeyBind__KEYBIND_AUTOROTATE

KeyBind__KEYBIND_AXIS_LINES

KeyBind__KEYBIND_BACK

KeyBind__KEYBIND_BREAK_LIQUIDS

KeyBind__KEYBIND_CHAT

KeyBind__KEYBIND_COUNT

KeyBind__KEYBIND_DELETE_BLOCK

KeyBind__KEYBIND_DROP_BLOCK

KeyBind__KEYBIND_EXT_INPUT

KeyBind__KEYBIND_FLY

KeyBind__KEYBIND_FLY_DOWN

KeyBind__KEYBIND_FLY_UP

KeyBind__KEYBIND_FOG

KeyBind__KEYBIND_FORWARD

KeyBind__KEYBIND_FULLSCREEN

KeyBind__KEYBIND_HALF_SPEED

KeyBind__KEYBIND_HIDE_FPS

KeyBind__KEYBIND_HIDE_GUI

KeyBind__KEYBIND_HOTBAR_SWITCH

KeyBind__KEYBIND_IDOVERLAY

KeyBind__KEYBIND_INVENTORY

KeyBind__KEYBIND_JUMP

KeyBind__KEYBIND_LEFT

KeyBind__KEYBIND_NOCLIP

KeyBind__KEYBIND_PICK_BLOCK

KeyBind__KEYBIND_PLACE_BLOCK

KeyBind__KEYBIND_PLAYER_LIST

KeyBind__KEYBIND_RESPAWN

KeyBind__KEYBIND_RIGHT

KeyBind__KEYBIND_SCREENSHOT

KeyBind__KEYBIND_SEND_CHAT

KeyBind__KEYBIND_SET_SPAWN

KeyBind__KEYBIND_SMOOTH_CAMERA

KeyBind__KEYBIND_SPEED

KeyBind__KEYBIND_THIRD_PERSON

KeyBind__KEYBIND_ZOOM_SCROLL

Key_Ascii_Names

Key_Function_Names

Key_INPUT_COUNT

Key_KEY_0

Key_KEY_1

Key_KEY_2

Key_KEY_3

Key_KEY_4

Key_KEY_5

Key_KEY_6

Key_KEY_7

Key_KEY_8

Key_KEY_9

Key_KEY_A

Key_KEY_B

Key_KEY_BACKSLASH

Key_KEY_BACKSPACE

Key_KEY_C

Key_KEY_CAPSLOCK

Key_KEY_COMMA

Key_KEY_D

Key_KEY_DELETE

Key_KEY_DOWN

Key_KEY_E

Key_KEY_END

Key_KEY_ENTER

Key_KEY_EQUALS

Key_KEY_ESCAPE

Key_KEY_F

Key_KEY_F1

Key_KEY_F2

Key_KEY_F3

Key_KEY_F4

Key_KEY_F5

Key_KEY_F6

Key_KEY_F7

Key_KEY_F8

Key_KEY_F9

Key_KEY_F10

Key_KEY_F11

Key_KEY_F12

Key_KEY_F13

Key_KEY_F14

Key_KEY_F15

Key_KEY_F16

Key_KEY_F17

Key_KEY_F18

Key_KEY_F19

Key_KEY_F20

Key_KEY_F21

Key_KEY_F22

Key_KEY_F23

Key_KEY_F24

Key_KEY_F25

Key_KEY_F26

Key_KEY_F27

Key_KEY_F28

Key_KEY_F29

Key_KEY_F30

Key_KEY_F31

Key_KEY_F32

Key_KEY_F33

Key_KEY_F34

Key_KEY_F35

Key_KEY_G

Key_KEY_H

Key_KEY_HOME

Key_KEY_I

Key_KEY_INSERT

Key_KEY_J

Key_KEY_K

Key_KEY_KP0

Key_KEY_KP1

Key_KEY_KP2

Key_KEY_KP3

Key_KEY_KP4

Key_KEY_KP5

Key_KEY_KP6

Key_KEY_KP7

Key_KEY_KP8

Key_KEY_KP9

Key_KEY_KP_DECIMAL

Key_KEY_KP_DIVIDE

Key_KEY_KP_ENTER

Key_KEY_KP_MINUS

Key_KEY_KP_MULTIPLY

Key_KEY_KP_PLUS

Key_KEY_L

Key_KEY_LALT

Key_KEY_LBRACKET

Key_KEY_LCTRL

Key_KEY_LEFT

Key_KEY_LMOUSE

Key_KEY_LSHIFT

Key_KEY_LWIN

Key_KEY_M

Key_KEY_MENU

Key_KEY_MINUS

Key_KEY_MMOUSE

Key_KEY_N

Key_KEY_NONE

Key_KEY_NUMLOCK

Key_KEY_O

Key_KEY_P

Key_KEY_PAGEDOWN

Key_KEY_PAGEUP

Key_KEY_PAUSE

Key_KEY_PERIOD

Key_KEY_PRINTSCREEN

Key_KEY_Q

Key_KEY_QUOTE

Key_KEY_R

Key_KEY_RALT

Key_KEY_RBRACKET

Key_KEY_RCTRL

Key_KEY_RIGHT

Key_KEY_RMOUSE

Key_KEY_RSHIFT

Key_KEY_RWIN

Key_KEY_S

Key_KEY_SCROLLLOCK

Key_KEY_SEMICOLON

Key_KEY_SLASH

Key_KEY_SPACE

Key_KEY_T

Key_KEY_TAB

Key_KEY_TILDE

Key_KEY_U

Key_KEY_UP

Key_KEY_V

Key_KEY_W

Key_KEY_X

Key_KEY_XBUTTON1

Key_KEY_XBUTTON2

Key_KEY_Y

Key_KEY_Z

LOCATIONUPDATE_PITCH

LOCATIONUPDATE_POS

LOCATIONUPDATE_ROTX

LOCATIONUPDATE_ROTZ

LOCATIONUPDATE_YAW

MATH_DEG2RAD

MATH_LARGENUM

MATH_PI

MATH_RAD2DEG

MENUINPUTWIDGET_MAX

MILLIS_PER_SEC

MODEL_BOX_VERTICES

MODEL_QUAD_VERTICES

MatrixType__MATRIX_PROJECTION

MatrixType__MATRIX_TEXTURE

MatrixType__MATRIX_VIEW

MsgType_MSG_TYPE_ANNOUNCEMENT

MsgType_MSG_TYPE_BOTTOMRIGHT_1

MsgType_MSG_TYPE_BOTTOMRIGHT_2

MsgType_MSG_TYPE_BOTTOMRIGHT_3

MsgType_MSG_TYPE_CLIENTSTATUS_1

MsgType_MSG_TYPE_CLIENTSTATUS_2

MsgType_MSG_TYPE_NORMAL

MsgType_MSG_TYPE_STATUS_1

MsgType_MSG_TYPE_STATUS_2

MsgType_MSG_TYPE_STATUS_3

NATIVE_STR_LEN

NameMode_NAME_MODE_ALL

NameMode_NAME_MODE_ALL_HOVERED

NameMode_NAME_MODE_ALL_UNSCALED

NameMode_NAME_MODE_COUNT

NameMode_NAME_MODE_HOVERED

NameMode_NAME_MODE_NONE

OGG_BUFFER_SIZE

OPCODE__OPCODE_ADD_ENTITY

OPCODE__OPCODE_BULK_BLOCK_UPDATE

OPCODE__OPCODE_COUNT

OPCODE__OPCODE_CUSTOM_BLOCK_LEVEL

OPCODE__OPCODE_DEFINE_BLOCK

OPCODE__OPCODE_DEFINE_BLOCK_EXT

OPCODE__OPCODE_DEFINE_EFFECT

OPCODE__OPCODE_ENTITY_TELEPORT

OPCODE__OPCODE_ENV_SET_COLOR

OPCODE__OPCODE_ENV_SET_MAP_APPEARANCE

OPCODE__OPCODE_ENV_SET_MAP_PROPERTY

OPCODE__OPCODE_ENV_SET_MAP_URL

OPCODE__OPCODE_ENV_SET_WEATHER

OPCODE__OPCODE_EXT_ADD_ENTITY

OPCODE__OPCODE_EXT_ADD_ENTITY2

OPCODE__OPCODE_EXT_ADD_PLAYER_NAME

OPCODE__OPCODE_EXT_ENTRY

OPCODE__OPCODE_EXT_INFO

OPCODE__OPCODE_EXT_REMOVE_PLAYER_NAME

OPCODE__OPCODE_HACK_CONTROL

OPCODE__OPCODE_HANDSHAKE

OPCODE__OPCODE_HOLD_THIS

OPCODE__OPCODE_KICK

OPCODE__OPCODE_LEVEL_BEGIN

OPCODE__OPCODE_LEVEL_DATA

OPCODE__OPCODE_LEVEL_END

OPCODE__OPCODE_MAKE_SELECTION

OPCODE__OPCODE_MESSAGE

OPCODE__OPCODE_ORI_UPDATE

OPCODE__OPCODE_PING

OPCODE__OPCODE_PLAYER_CLICK

OPCODE__OPCODE_RELPOS_AND_ORI_UPDATE

OPCODE__OPCODE_RELPOS_UPDATE

OPCODE__OPCODE_REMOVE_ENTITY

OPCODE__OPCODE_REMOVE_SELECTION

OPCODE__OPCODE_SET_BLOCK

OPCODE__OPCODE_SET_BLOCK_CLIENT

OPCODE__OPCODE_SET_BLOCK_PERMISSION

OPCODE__OPCODE_SET_ENTITY_PROPERTY

OPCODE__OPCODE_SET_HOTBAR

OPCODE__OPCODE_SET_INVENTORY_ORDER

OPCODE__OPCODE_SET_MODEL

OPCODE__OPCODE_SET_PERMISSION

OPCODE__OPCODE_SET_REACH

OPCODE__OPCODE_SET_SPAWNPOINT

OPCODE__OPCODE_SET_TEXT_COLOR

OPCODE__OPCODE_SET_TEXT_HOTKEY

OPCODE__OPCODE_SPAWN_EFFECT

OPCODE__OPCODE_TWO_WAY_PING

OPCODE__OPCODE_UNDEFINE_BLOCK

OPCODE__OPCODE_VELOCITY_CONTROL

OPT_AUTO_CLOSE_LAUNCHER

OPT_BLACK_TEXT

OPT_BLOCK_PHYSICS

OPT_CAMERA_CLIPPING

OPT_CHATLINES

OPT_CHAT_LOGGING

OPT_CHAT_SCALE

OPT_CLASSIC_ARM_MODEL

OPT_CLASSIC_CHAT

OPT_CLASSIC_GUI

OPT_CLASSIC_HACKS

OPT_CLASSIC_MODE

OPT_CLASSIC_OPTIONS

OPT_CLASSIC_TABLIST

OPT_CLICKABLE_CHAT

OPT_CPE

OPT_CUSTOM_BLOCKS

OPT_DEFAULT_TEX_PACK

OPT_ENTITY_SHADOW

OPT_FIELD_OF_VIEW

OPT_FONT_NAME

OPT_FORCE_OPENAL

OPT_FPS_LIMIT

OPT_FULL_BLOCK_STEP

OPT_HACKS_ENABLED

OPT_HACK_PERM_MSGS

OPT_HOTBAR_SCALE

OPT_INVENTORY_SCALE

OPT_INVERT_MOUSE

OPT_JUMP_VELOCITY

OPT_MAX_CHUNK_UPDATES

OPT_MAX_MUSIC_DELAY

OPT_MIN_MUSIC_DELAY

OPT_MIPMAPS

OPT_MODIFIABLE_LIQUIDS

OPT_MUSIC_VOLUME

OPT_NAMES_MODE

OPT_NOCLIP_SLIDE

OPT_PUSHBACK_PLACING

OPT_RENDER_TYPE

OPT_SENSITIVITY

OPT_SERVER_TEXTURES

OPT_SHOW_BLOCK_IN_HAND

OPT_SHOW_FPS

OPT_SIMPLE_ARMS_ANIM

OPT_SMOOTH_LIGHTING

OPT_SOUND_VOLUME

OPT_SPEED_FACTOR

OPT_SURVIVAL_MODE

OPT_TAB_AUTOCOMPLETE

OPT_USE_CHAT_FONT

OPT_USE_MUSIC

OPT_USE_SOUND

OPT_VIEW_BOBBING

OPT_VIEW_DISTANCE

OPT_WINDOW_HEIGHT

OPT_WINDOW_WIDTH

OPT_WOM_STYLE_HACKS

PACKEDCOL_A_MASK

PACKEDCOL_A_SHIFT

PACKEDCOL_B_MASK

PACKEDCOL_B_SHIFT

PACKEDCOL_G_MASK

PACKEDCOL_G_SHIFT

PACKEDCOL_RGB_MASK

PACKEDCOL_R_MASK

PACKEDCOL_R_SHIFT

PACKEDCOL_SHADE_X

PACKEDCOL_SHADE_YMIN

PACKEDCOL_SHADE_Z

PACKEDCOL_WHITE

PNG_MAX_DIMS

Pointers_Count

RESPAWN_NOT_FOUND

RotateOrder_ROTATE_ORDER_XZY

RotateOrder_ROTATE_ORDER_YZX

RotateOrder_ROTATE_ORDER_ZYX

SECS_PER_DAY

SECS_PER_HOUR

SECS_PER_MIN

SKIN_FETCH_COMPLETED

SKIN_FETCH_DOWNLOADING

SKIN_TYPE_SKIN_64x32

SKIN_TYPE_SKIN_64x64

SKIN_TYPE_SKIN_64x64_SLIM

SKIN_TYPE_SKIN_INVALID

STRINGSBUFFER_BUFFER_DEF_SIZE

STRINGSBUFFER_FLAGS_DEF_ELEMS

STRINGSBUFFER_LEN_MASK

STRINGSBUFFER_LEN_SHIFT

STRING_INT_CHARS

STRING_SIZE

ShadowMode_SHADOW_MODE_CIRCLE

ShadowMode_SHADOW_MODE_CIRCLE_ALL

ShadowMode_SHADOW_MODE_COUNT

ShadowMode_SHADOW_MODE_NONE

ShadowMode_SHADOW_MODE_SNAP_TO_BLOCK

Socket_PollMode_SOCKET_POLL_READ

Socket_PollMode_SOCKET_POLL_WRITE

SoundType_SOUND_CLOTH

SoundType_SOUND_COUNT

SoundType_SOUND_GLASS

SoundType_SOUND_GRASS

SoundType_SOUND_GRAVEL

SoundType_SOUND_METAL

SoundType_SOUND_NONE

SoundType_SOUND_SAND

SoundType_SOUND_SNOW

SoundType_SOUND_STONE

SoundType_SOUND_WOOD

TABLIST_MAX_NAMES

TEXTATLAS_MAX_WIDTHS

TEXTGROUPWIDGET_LEN

TEXTGROUPWIDGET_MAX_LINES

TEXTWIDGET_MAX

TREE_MAX_COUNT

UNIX_EPOCH

UPDATE_FILE

URL_MAX_SIZE

UV2_Scale

VORBIS_MAX_BLOCK_SIZE

VORBIS_MAX_CHANS

VertexFormat__VERTEX_FORMAT_P3FC4B

VertexFormat__VERTEX_FORMAT_P3FT2FC4B

Weather__WEATHER_RAINY

Weather__WEATHER_SNOWY

Weather__WEATHER_SUNNY

WindowState_WINDOW_STATE_FULLSCREEN

WindowState_WINDOW_STATE_MINIMISED

WindowState_WINDOW_STATE_NORMAL

ZIP_MAX_ENTRIES

_NL

keybindNames

Statics

Atlas1D

Atlas2D

BlockEvents

Blocks

ChatEvents

Drawer

Entities

EntityEvents

Env

Game

Gfx

GfxEvents

InputEvents

Inventory

Models

NetEvents

Physics

PointerEvents

Server

TabList

TabListEvents

TextureEvents

UserEvents

WindowEvents

World

WorldEvents

Functions

AABB_Make⚠

Bitmap_Scale⚠

Block_FindID⚠

Block_Parse⚠

Block_UNSAFE_GetName⚠

BoxDesc_BuildBox⚠

BoxDesc_BuildRotatedBox⚠

BoxDesc_XQuad⚠

BoxDesc_YQuad⚠

BoxDesc_ZQuad⚠

Camera_Register⚠

Chat_Add⚠

Chat_AddOf⚠

Chat_AddRaw⚠

Chat_Send⚠

Clipboard_GetText⚠

Clipboard_SetText⚠

Commands_Register⚠

Convert_ParseBool⚠

Convert_ParseFloat⚠

Convert_ParseInt⚠

Convert_ParseUInt8⚠

Convert_ParseUInt16⚠

Convert_ParseUInt64⚠

DateTime_CurrentLocal⚠

DateTime_CurrentUTC_MS⚠

Deflate_MakeStream⚠

Directory_Create⚠

Directory_Enum⚠

Directory_Exists⚠

Drawer2D_BmpCopy⚠

Drawer2D_Clear⚠

Drawer2D_DrawText⚠

Drawer2D_MeasureText⚠

Drawer2D_MakeTextTexture⚠

Drawer2D_MakeTexture⚠

Drawer_XMax⚠

Drawer_XMin⚠

Drawer_YMax⚠

Drawer_YMin⚠

Drawer_ZMax⚠

Drawer_ZMin⚠

DynamicLib_Get⚠

DynamicLib_Load⚠

Entity_GetTransform⚠

Entity_SetModel⚠

Entity_TouchesAny⚠

Env_Reset⚠

Env_SetCloudsCol⚠

Env_SetCloudsHeight⚠

Env_SetCloudsSpeed⚠

Env_SetEdgeBlock⚠

Env_SetEdgeHeight⚠

Env_SetExpFog⚠

Env_SetFogCol⚠

Env_SetShadowCol⚠

Env_SetSidesBlock⚠

Env_SetSidesOffset⚠

Env_SetSkyCol⚠

Env_SetSkyboxCol⚠

Env_SetSkyboxHorSpeed⚠

Env_SetSkyboxVerSpeed⚠

Env_SetSunCol⚠

Env_SetWeather⚠

Env_SetWeatherFade⚠

Env_SetWeatherSpeed⚠

Event_RaiseBlock⚠

Event_RaiseChat⚠

Event_RaiseEntry⚠

Event_RaiseFloat⚠

Event_RaiseInput⚠

Event_RaiseInt⚠

Event_RaiseMove⚠

Event_RaiseString⚠

Event_RaiseVoid⚠

Event_Register⚠

Event_RegisterBlock⚠

Event_RegisterChat⚠

Event_RegisterEntry⚠

Event_RegisterFloat⚠

Event_RegisterInput⚠

Event_RegisterInt⚠

Event_RegisterMove⚠

Event_RegisterString⚠

Event_RegisterVoid⚠

Event_Unregister⚠

Event_UnregisterBlock⚠

Event_UnregisterChat⚠

Event_UnregisterEntry⚠

Event_UnregisterFloat⚠

Event_UnregisterInput⚠

Event_UnregisterInt⚠

Event_UnregisterMove⚠

Event_UnregisterString⚠

Event_UnregisterVoid⚠

File_Exists⚠

File_SetModifiedTime⚠

Font_Free⚠

GZip_MakeStream⚠

Game_ChangeBlock⚠

Game_UpdateBlock⚠

Gfx_BindIb⚠

Gfx_BindTexture⚠

Gfx_BindVb⚠

Gfx_Clear⚠

Gfx_ClearCol⚠

Gfx_CreateDynamicVb⚠

Gfx_CreateIb⚠

Gfx_CreateTexture⚠

Gfx_CreateVb⚠

Gfx_DeleteIb⚠

Gfx_DeleteTexture⚠

Gfx_DeleteVb⚠

Gfx_DisableMipmaps⚠

Gfx_DrawVb_IndexedTris⚠

Gfx_DrawVb_IndexedTris_Range⚠

Gfx_DrawVb_Lines⚠

Gfx_EnableMipmaps⚠

Gfx_GetFog⚠

Gfx_LoadIdentityMatrix⚠

Gfx_LoadMatrix⚠

Gfx_LockDynamicVb⚠

Gfx_LockVb⚠

Gfx_SetAlphaArgBlend⚠

Gfx_SetAlphaBlending⚠

Gfx_SetAlphaTest⚠

Gfx_SetColWriteMask⚠

Gfx_SetDepthTest⚠

Gfx_SetDepthWrite⚠

Gfx_SetDynamicVbData⚠

Gfx_SetFaceCulling⚠

Gfx_SetFog⚠

Gfx_SetFogCol⚠

Gfx_SetFogDensity⚠

Gfx_SetFogEnd⚠

Gfx_SetFogMode⚠

Gfx_SetTexturing⚠

Gfx_SetVertexFormat⚠

Gfx_UnlockDynamicVb⚠

Gfx_UnlockVb⚠

Gfx_UpdateTexturePart⚠

Gradient_Blend⚠

Gradient_Noise⚠

Gradient_Tint⚠

Gradient_Vertical⚠

Inflate_Init⚠

Inflate_MakeStream⚠

Map_FindImporter⚠

Map_LoadFrom⚠

Math_Cos⚠

Math_Sin⚠

Matrix_Mul⚠

Matrix_RotateX⚠

Matrix_RotateY⚠

Matrix_RotateZ⚠

Matrix_Scale⚠

Matrix_Translate⚠

Mem_Alloc⚠

Mem_AllocCleared⚠

Mem_Free⚠

Mem_Realloc⚠

Mem_TryAlloc⚠

Mem_TryAllocCleared⚠

Mem_TryRealloc⚠

Model_ApplyTexture⚠

Model_DrawArmPart⚠

Model_DrawPart⚠

Model_DrawRotate⚠

Model_Get⚠

Model_GetTexture⚠

Model_Init⚠

Model_Register⚠

Model_RegisterTexture⚠

Model_Render⚠

Model_SetupState⚠

Model_UpdateVB⚠

Mutex_Create⚠

Mutex_Free⚠

Mutex_Lock⚠

Mutex_Unlock⚠

Options_Get⚠

Options_GetBool⚠

Options_GetEnum⚠

Options_GetFloat⚠

Options_GetInt⚠

Options_Load⚠

Options_Save⚠

Options_Set⚠

Options_SetBool⚠

Options_SetInt⚠

Options_SetString⚠

PackedCol_A_Bits

PackedCol_B_Bits

PackedCol_G_Bits

PackedCol_Lerp⚠

PackedCol_Make

PackedCol_R_Bits

PackedCol_Scale⚠

PackedCol_Tint⚠

Platform_ConvertString⚠

Png_Decode⚠

Png_Encode⚠

Process_Exit⚠

Process_StartGame⚠

Process_StartOpen⚠

Random_Float⚠

Random_Next⚠

Random_Seed⚠

Socket_Available⚠

Socket_Close⚠

Socket_Connect⚠

Socket_Create⚠

Socket_GetError⚠

Socket_Poll⚠

Socket_Read⚠

Socket_SetBlocking⚠

Socket_Write⚠

Stopwatch_ElapsedMicroseconds⚠

Stopwatch_Measure⚠

Stream_CreateFile⚠

Stream_FromFile⚠

Stream_OpenFile⚠

Stream_Read⚠

Stream_ReadLine⚠

Stream_ReadonlyBuffered⚠

Stream_ReadonlyMemory⚠

Stream_ReadonlyPortion⚠

Stream_Write⚠

Stream_WriteLine⚠

String_Append⚠

String_AppendBool⚠

String_AppendColorless⚠

String_AppendConst⚠

String_AppendFloat⚠

String_AppendHex⚠

String_AppendInt⚠

String_AppendPaddedInt⚠

String_AppendString⚠

String_AppendUInt32⚠

String_CalcLen⚠

String_CaselessContains⚠

String_CaselessEnds⚠

String_CaselessEquals⚠

String_CaselessEqualsConst⚠

String_CaselessStarts⚠

String_Compare⚠

String_Copy⚠

String_CopyToRaw⚠

String_DeleteAt⚠

String_Equals⚠

String_Format1⚠

String_Format2⚠

String_Format3⚠

String_Format4⚠

String_FromReadonly⚠

String_IndexOfAt⚠

String_IndexOfConst⚠

String_Init⚠

String_InsertAt⚠

String_LastIndexOfAt⚠

String_StripCols⚠

String_UNSAFE_Separate⚠

String_UNSAFE_Split⚠

String_UNSAFE_SplitBy⚠

String_UNSAFE_Substring⚠

String_UNSAFE_SubstringAt⚠

String_UNSAFE_TrimEnd⚠

String_UNSAFE_TrimStart⚠

StringsBuffer_Add⚠

StringsBuffer_Clear⚠

StringsBuffer_Remove⚠

StringsBuffer_UNSAFE_Get⚠

TabList_Remove⚠

TabList_Set⚠

Thread_Detach⚠

Thread_Join⚠

Thread_Sleep⚠

Thread_Start⚠

Waitable_Create⚠

Waitable_Free⚠

Waitable_Signal⚠

Waitable_Wait⚠

Waitable_WaitFor⚠

Window_SetTitle⚠

Window_ShowDialog⚠

World_ApplyTexturePack⚠

World_Blocks

World_Blocks2

World_GetBlock

World_NewMap⚠

World_Pack

World_SetNewMap⚠

ZLib_MakeStream⚠

Zip_Extract⚠

Zip_Init⚠

Type Definitions

AudioHandle

BLOCKID

Bitmap

BitmapCol

BlockID

BlockRaw

Button_Get

Button_Set

Codepoint

CollideType

Directory_EnumCallback

DrawType

ERRORS_ALL

EntityID

EntityType

Entity_TouchesCondition

EntryList_Filter

EnvVar

Event_Block_Callback

Event_Chat_Callback

Event_Entry_Callback

Event_Float_Callback

Event_Input_Callback

Event_Int_Callback

Event_PointerMove_Callback

Event_String_Callback

Event_Void_Callback

FACE_CONSTS

FONT_STYLE

Face

FileHandle

File_SeekFrom

FogFunc

FogFunc_

FpsLimitMethod

GfxResourceID

GuiAnchor

GuiPriority

HotkeyModifiers

HttpProgress

HttpRequestType

IMapImporter

IVec3

JsonOnNew

JsonOnValue

Key

KeyBind

KeyBind_

LScreen_Func

LWidget_Func

Logger_DescribeError

Logger_DoWarn

MatrixType

MatrixType_

MsgType

NameMode

Net_Handler

OPCODE_

PackedCol

PhysicsHandler

Png_RowSelector

RNGState

Rect2D

RequestClipboardCallback

RotateOrder

SKIN_TYPE

ScheduledTaskCallback

ShadowMode

Size2D

SocketHandle

Socket_PollMode

SoundType

SpecialInputAppendFunc

String

StringsBuffer

TextGroupWidget_Get

TextureLoc

TextureRec

Thread_StartFunc

TimeMS

Vec2

Vec3

VertexFormat

VertexFormat_

VertexP3fC4b

VertexP3fT2fC4b

Weather_

Widget_LeftClick

WindowState

cc_bool

cc_int8

cc_int16

cc_int32

cc_int64

cc_result

cc_uint8

cc_uint16

cc_uint32

cc_uint64

cc_uintptr

Unions

InflateState__bindgen_ty_1

IntAndFloat

MenuInputDesc__bindgen_ty_1

Stream__bindgen_ty_1