Crate classicube_sys

classicube-sys

Rust bindings to ClassiCube

Documentation

Example

Add this to Cargo.toml:

[lib]
crate-type = ["cdylib"]

[dependencies]
classicube-sys = { git = "https://github.com/SpiralP/rust-classicube-sys.git" }

lib.rs:

use classicube_sys::*;
use std::{os::raw::c_int, ptr};

extern "C" fn init() {
    let owned_string = OwnedString::new("hello from rust!");
    unsafe {
        Chat_Add(owned_string.as_cc_string());
    }
}

#[no_mangle]
pub static Plugin_ApiVersion: c_int = 1;

#[no_mangle]
pub static mut Plugin_Component: IGameComponent = IGameComponent {
    // Called when the game is being loaded.
    Init: Some(init),
    // Called when the component is being freed. (e.g. due to game being closed)
    Free: None,
    // Called to reset the component's state. (e.g. reconnecting to server)
    Reset: None,
    // Called to update the component's state when the user begins loading a new map.
    OnNewMap: None,
    // Called to update the component's state when the user has finished loading a new map.
    OnNewMapLoaded: None,
    // Next component in linked list of components.
    next: ptr::null_mut(),
};

Copy the .dll/.so/.dylib from target/{debug,release}/ to the plugins folder where your ClassiCube executable is found

References

Example library usage

ClassiCube source

ClassiCube Plugin help

ClassiCube Plugin examples

Modules

screen

Macros

BoxDesc_Bounds

gives (x1, y1, z1, x2, y2, z2)

BoxDesc_Box

gives ((x, y, z), (x1, y1, z1, x2, y2, z2))

BoxDesc_Dim

BoxDesc_Dims

gives (x, y, z)

BoxDesc_Rot

gives (x, y, z)

BoxDesc_Tex

Model_RetAABB

Model_RetSize

Vec3_AddBy

Vec3_IsZero

Whether all components of a 3D vector are 0

Vec3_Mul1By

Vec3_Mul3By

Vec3_Set

Sets the X, Y, and Z components of a 3D vector

Vec3_SubBy

Structs

AABB

AnimatedComp

AudioContext

AudioData

Bitmap

BoxDesc

ButtonWidget

Camera

ChatCommand

ChatInputWidget

CheckUpdateData

ChunkInfo

ChunkPartInfo

Codebook

CollisionsComp

CustomModel

CustomModelAnim

CustomModelPart

CustomModelPartDef

CustomParticleEffect

DateTime

DeflateState

DrawTextArgs

DynamicLibSym

Entity

EntityVTABLE

Event_Block

Event_Chat

Event_Entry

Event_Float

Event_Input

Event_Int

Event_PluginMessage

Event_RawMove

Event_String

Event_Void

FetchFlagsData

FetchServerData

FetchServersData

FetchUpdateData

Floor

FontDesc

GZipHeader

GZipState

GetTokenTaskData

HacksComp

HotbarWidget

HotkeyData

HttpRequest

HuffmanTable

IGameComponent

IVec3_

InflateState

InputWidget

InterpComp

InterpState

JsonContext

LButton

LCheckbox

LInput

LLabel

LLine

LScreen

LSlider

LTable

LTableCell

LTableColumn

LWebTask

LWidget

LWidgetVTABLE

LauncherTheme

LocalPlayer

LocalPlayerInput

LocationUpdate

Mapping

Matrix

MenuInputDesc

MenuInputDesc__bindgen_ty_1__bindgen_ty_1

MenuInputDesc__bindgen_ty_1__bindgen_ty_2

MenuInputDesc__bindgen_ty_1__bindgen_ty_3

MenuInputDesc__bindgen_ty_1__bindgen_ty_4

MenuInputVTABLE

Mode

Model

ModelPart

ModelTex

ModelVertex

NetInterpComp

NetPlayer

OggState

OpenKeyboardArgs

OwnedChatCommand

OwnedGfxTexture

OwnedGfxVertexBuffer

OwnedString

Particle

PhysicsComp

Physics_

Pointer

RayTracer

Rect2D_

Residue

ScheduledTask

Screen

ScreenVTABLE

ScrollbarWidget

SearcherState

ServerInfo

SignInTaskData

SpecialInputTab

SpecialInputWidget

Stream

Stream__bindgen_ty_1__bindgen_ty_1

Stream__bindgen_ty_1__bindgen_ty_2

Stream__bindgen_ty_1__bindgen_ty_3

Stream__bindgen_ty_1__bindgen_ty_4

StringsBuffer

TableWidget

TextAtlas

TextGroupWidget

TextInputWidget

TextWidget

Texture

TextureRec_

TiltComp

Vec2_

Vec3_

Vec4

VertexColoured

VertexTextured

VorbisState

VorbisWindow

Widget

WidgetVTABLE

ZLibHeader

ZLibState

ZipEntry

ZipState

_Atlas1DData

_Atlas2DData

_BlockEventsList

_BlockLists

_CameraData

_ChatEventsList

_DisplayData

_Drawer2DData

_DrawerData

_EntitiesData

_EntityEventsList

_EnvData

_GameData

_GfxData

_GfxEventsList

_GuiData

_InputEventsList

_InventoryData

_ModelsData

_NetEventsList

_PluginMessageEventsList

_PointerEventsList

_ProtocolData

_ServerConnectionData

_TabListData

_TabListEventsList

_TextureEventsList

_UserEventsList

_WinData

_WindowEventsList

_WorldData

_WorldEventsList

__BindgenBitfieldUnit

cc_string_

imdct_state

Constants

ATLAS1D_MAX_ATLASES

ATLAS2D_MASK

ATLAS2D_MAX_ROWS_COUNT

ATLAS2D_SHIFT

ATLAS2D_TILES_PER_ROW

AUDIO_MAX_BUFFERS

BITMAPCOL_A_MASK

BITMAPCOL_A_SHIFT

BITMAPCOL_B_MASK

BITMAPCOL_B_SHIFT

BITMAPCOL_G_MASK

BITMAPCOL_G_SHIFT

BITMAPCOL_RGB_MASK

BITMAPCOL_R_MASK

BITMAPCOL_R_SHIFT

BLOCKID_BLOCK_AIR

BLOCKID_BLOCK_AQUA

BLOCKID_BLOCK_BEDROCK

BLOCKID_BLOCK_BLACK

BLOCKID_BLOCK_BLUE

BLOCKID_BLOCK_BOOKSHELF

BLOCKID_BLOCK_BRICK

BLOCKID_BLOCK_BROWN

BLOCKID_BLOCK_BROWN_SHROOM

BLOCKID_BLOCK_CERAMIC_TILE

BLOCKID_BLOCK_COAL_ORE

BLOCKID_BLOCK_COBBLE

BLOCKID_BLOCK_COBBLE_SLAB

BLOCKID_BLOCK_COUNT

BLOCKID_BLOCK_CPE_COUNT

BLOCKID_BLOCK_CRATE

BLOCKID_BLOCK_CYAN

BLOCKID_BLOCK_DANDELION

BLOCKID_BLOCK_DEEP_BLUE

BLOCKID_BLOCK_DIRT

BLOCKID_BLOCK_DOUBLE_SLAB

BLOCKID_BLOCK_FIRE

BLOCKID_BLOCK_FOREST_GREEN

BLOCKID_BLOCK_GLASS

BLOCKID_BLOCK_GOLD

BLOCKID_BLOCK_GOLD_ORE

BLOCKID_BLOCK_GRASS

BLOCKID_BLOCK_GRAVEL

BLOCKID_BLOCK_GRAY

BLOCKID_BLOCK_GREEN

BLOCKID_BLOCK_ICE

BLOCKID_BLOCK_INDIGO

BLOCKID_BLOCK_IRON

BLOCKID_BLOCK_IRON_ORE

BLOCKID_BLOCK_LAVA

BLOCKID_BLOCK_LEAVES

BLOCKID_BLOCK_LIGHT_PINK

BLOCKID_BLOCK_LIME

BLOCKID_BLOCK_LOG

BLOCKID_BLOCK_MAGENTA

BLOCKID_BLOCK_MAGMA

BLOCKID_BLOCK_MAX_CPE

BLOCKID_BLOCK_MAX_DEFINED

BLOCKID_BLOCK_MAX_ORIGINAL

BLOCKID_BLOCK_MOSSY_ROCKS

BLOCKID_BLOCK_OBSIDIAN

BLOCKID_BLOCK_ORANGE

BLOCKID_BLOCK_ORIGINAL_COUNT

BLOCKID_BLOCK_PILLAR

BLOCKID_BLOCK_PINK

BLOCKID_BLOCK_RED

BLOCKID_BLOCK_RED_SHROOM

BLOCKID_BLOCK_ROPE

BLOCKID_BLOCK_ROSE

BLOCKID_BLOCK_SAND

BLOCKID_BLOCK_SANDSTONE

BLOCKID_BLOCK_SAPLING

BLOCKID_BLOCK_SLAB

BLOCKID_BLOCK_SNOW

BLOCKID_BLOCK_SPONGE

BLOCKID_BLOCK_STILL_LAVA

BLOCKID_BLOCK_STILL_WATER

BLOCKID_BLOCK_STONE

BLOCKID_BLOCK_STONE_BRICK

BLOCKID_BLOCK_TEAL

BLOCKID_BLOCK_TNT

BLOCKID_BLOCK_TURQUOISE

BLOCKID_BLOCK_VIOLET

BLOCKID_BLOCK_WATER

BLOCKID_BLOCK_WHITE

BLOCKID_BLOCK_WOOD

BLOCKID_BLOCK_YELLOW

BUTTONWIDGET_MAX

CC_ERRORS_AL_ERR_INIT_CONTEXT

CC_ERRORS_AL_ERR_INIT_DEVICE

CC_ERRORS_CW_ERR_ROOT_TAG

CC_ERRORS_CW_ERR_STRING_LEN

CC_ERRORS_CW_ERR_UUID_LEN

CC_ERRORS_DAT_ERR_IDENTIFIER

CC_ERRORS_DAT_ERR_JARRAY_CONTENT

CC_ERRORS_DAT_ERR_JARRAY_TYPE

CC_ERRORS_DAT_ERR_JCLASS_ANNOTATION

CC_ERRORS_DAT_ERR_JCLASS_FIELDS

CC_ERRORS_DAT_ERR_JCLASS_TYPE

CC_ERRORS_DAT_ERR_JFIELD_CLASS_NAME

CC_ERRORS_DAT_ERR_JIDENTIFIER

CC_ERRORS_DAT_ERR_JOBJECT_TYPE

CC_ERRORS_DAT_ERR_JSTRING_LEN

CC_ERRORS_DAT_ERR_JVERSION

CC_ERRORS_DAT_ERR_ROOT_TYPE

CC_ERRORS_DAT_ERR_VERSION

CC_ERRORS_ERROR_BASE

CC_ERRORS_ERR_DOWNLOAD_INVALID

CC_ERRORS_ERR_END_OF_STREAM

CC_ERRORS_ERR_INVALID_ARGUMENT

CC_ERRORS_ERR_NOT_SUPPORTED

CC_ERRORS_ERR_NO_AUDIO_OUTPUT

CC_ERRORS_ERR_OUT_OF_MEMORY

CC_ERRORS_FCM_ERR_IDENTIFIER

CC_ERRORS_FCM_ERR_REVISION

CC_ERRORS_GZIP_ERR_FLAGS

CC_ERRORS_GZIP_ERR_HEADER1

CC_ERRORS_GZIP_ERR_HEADER2

CC_ERRORS_GZIP_ERR_METHOD

CC_ERRORS_INF_ERR_BLOCKTYPE

CC_ERRORS_INF_ERR_INVALID_CODE

CC_ERRORS_INF_ERR_LEN_VERIFY

CC_ERRORS_INF_ERR_NUM_CODES

CC_ERRORS_INF_ERR_REPEAT_BEG

CC_ERRORS_INF_ERR_REPEAT_END

CC_ERRORS_LVL_ERR_VERSION

CC_ERRORS_NBT_ERR_UNKNOWN

CC_ERRORS_OGG_ERR_INVALID_SIG

CC_ERRORS_OGG_ERR_VERSION

CC_ERRORS_PNG_ERR_COMP_METHOD

CC_ERRORS_PNG_ERR_FILTER

CC_ERRORS_PNG_ERR_INTERLACED

CC_ERRORS_PNG_ERR_INVALID_COL_BPP

CC_ERRORS_PNG_ERR_INVALID_HDR_SIZE

CC_ERRORS_PNG_ERR_INVALID_SCANLINE

CC_ERRORS_PNG_ERR_INVALID_SIG

CC_ERRORS_PNG_ERR_NO_DATA

CC_ERRORS_PNG_ERR_PAL_SIZE

CC_ERRORS_PNG_ERR_REACHED_IEND

CC_ERRORS_PNG_ERR_TOO_TALL

CC_ERRORS_PNG_ERR_TOO_WIDE

CC_ERRORS_PNG_ERR_TRANS_COUNT

CC_ERRORS_PNG_ERR_TRANS_INVALID

CC_ERRORS_VORBIS_ERR_BLOCKSIZE

CC_ERRORS_VORBIS_ERR_CHANS

CC_ERRORS_VORBIS_ERR_CODEBOOK_ENTRY

CC_ERRORS_VORBIS_ERR_CODEBOOK_LOOKUP

CC_ERRORS_VORBIS_ERR_CODEBOOK_SYNC

CC_ERRORS_VORBIS_ERR_FLOOR_TYPE

CC_ERRORS_VORBIS_ERR_FRAME_TYPE

CC_ERRORS_VORBIS_ERR_FRAMING

CC_ERRORS_VORBIS_ERR_MAPPING_CHANS

CC_ERRORS_VORBIS_ERR_MAPPING_RESERVED

CC_ERRORS_VORBIS_ERR_MAPPING_TYPE

CC_ERRORS_VORBIS_ERR_MODE_TRANSFORM

CC_ERRORS_VORBIS_ERR_MODE_TYPE

CC_ERRORS_VORBIS_ERR_MODE_WINDOW

CC_ERRORS_VORBIS_ERR_RESIDUE_TYPE

CC_ERRORS_VORBIS_ERR_TIME_TYPE

CC_ERRORS_VORBIS_ERR_VERSION

CC_ERRORS_VORBIS_ERR_WRONG_HEADER

CC_ERRORS_WAV_ERR_DATA_TYPE

CC_ERRORS_WAV_ERR_SAMPLE_BITS

CC_ERRORS_WAV_ERR_STREAM_HDR

CC_ERRORS_WAV_ERR_STREAM_TYPE

CC_ERRORS_ZIP_ERR_FILENAME_LEN

CC_ERRORS_ZIP_ERR_INVALID_CENTRAL_DIR

CC_ERRORS_ZIP_ERR_INVALID_LOCAL_DIR

CC_ERRORS_ZIP_ERR_NO_END_OF_CENTRAL_DIR

CC_ERRORS_ZIP_ERR_SEEK_CENTRAL_DIR

CC_ERRORS_ZIP_ERR_SEEK_END_OF_CENTRAL_DIR

CC_ERRORS_ZIP_ERR_SEEK_LOCAL_DIR

CC_ERRORS_ZIP_ERR_TOO_MANY_ENTRIES

CC_ERRORS_ZLIB_ERR_FLAGS

CC_ERRORS_ZLIB_ERR_METHOD

CHUNK_MASK

CHUNK_MAX

CHUNK_SHIFT

CHUNK_SIZE

CHUNK_SIZE_2

CHUNK_SIZE_3

CollideType_COLLIDE_CLIMB_ROPE

CollideType_COLLIDE_GAS

CollideType_COLLIDE_ICE

CollideType_COLLIDE_LIQUID

CollideType_COLLIDE_LIQUID_LAVA

CollideType_COLLIDE_LIQUID_WATER

CollideType_COLLIDE_SLIPPERY_ICE

CollideType_COLLIDE_SOLID

CustomModelAnimAxis_CustomModelAnimAxis_X

CustomModelAnimAxis_CustomModelAnimAxis_Y

CustomModelAnimAxis_CustomModelAnimAxis_Z

CustomModelAnimType_CustomModelAnimType_FlipRotate

CustomModelAnimType_CustomModelAnimType_FlipRotateVelocity

CustomModelAnimType_CustomModelAnimType_FlipSize

CustomModelAnimType_CustomModelAnimType_FlipSizeVelocity

CustomModelAnimType_CustomModelAnimType_FlipTranslate

CustomModelAnimType_CustomModelAnimType_FlipTranslateVelocity

CustomModelAnimType_CustomModelAnimType_Head

CustomModelAnimType_CustomModelAnimType_LeftArmX

CustomModelAnimType_CustomModelAnimType_LeftArmZ

CustomModelAnimType_CustomModelAnimType_LeftLegX

CustomModelAnimType_CustomModelAnimType_None

CustomModelAnimType_CustomModelAnimType_RightArmX

CustomModelAnimType_CustomModelAnimType_RightArmZ

CustomModelAnimType_CustomModelAnimType_RightLegX

CustomModelAnimType_CustomModelAnimType_SinRotate

CustomModelAnimType_CustomModelAnimType_SinRotateVelocity

CustomModelAnimType_CustomModelAnimType_SinSize

CustomModelAnimType_CustomModelAnimType_SinSizeVelocity

CustomModelAnimType_CustomModelAnimType_SinTranslate

CustomModelAnimType_CustomModelAnimType_SinTranslateVelocity

CustomModelAnimType_CustomModelAnimType_Spin

CustomModelAnimType_CustomModelAnimType_SpinVelocity

DEFAULT_MAX_VIEWDIST

DEFAULT_MUSIC_VOLUME

DEFAULT_SOUNDS_VOLUME

DEFLATE_BLOCK_SIZE

DEFLATE_BUFFER_SIZE

DEFLATE_HASH_MASK

DEFLATE_HASH_SIZE

DEFLATE_OUT_SIZE

DRAWER2D_MAX_COLORS

DrawType_DRAW_GAS

DrawType_DRAW_OPAQUE

DrawType_DRAW_SPRITE

DrawType_DRAW_TRANSLUCENT

DrawType_DRAW_TRANSPARENT

DrawType_DRAW_TRANSPARENT_THICK

ENTITIES_MAX_COUNT

ENTITIES_SELF_ID

ENTITY_ADJUSTMENT

ENV_LEGACY

ENV_MINIMAL

EVENT_MAX_CALLBACKS

EXTCHUNK_SIZE

EXTCHUNK_SIZE_2

EXTCHUNK_SIZE_3

EntityType_ENTITY_TYPE_NONE

EntityType_ENTITY_TYPE_PLAYER

EnvVar_ENV_VAR_CLOUDS_COLOR

EnvVar_ENV_VAR_CLOUDS_HEIGHT

EnvVar_ENV_VAR_CLOUDS_SPEED

EnvVar_ENV_VAR_EDGE_BLOCK

EnvVar_ENV_VAR_EDGE_HEIGHT

EnvVar_ENV_VAR_EXP_FOG

EnvVar_ENV_VAR_FOG_COLOR

EnvVar_ENV_VAR_SHADOW_COLOR

EnvVar_ENV_VAR_SIDES_BLOCK

EnvVar_ENV_VAR_SIDES_OFFSET

EnvVar_ENV_VAR_SKYBOX_COLOR

EnvVar_ENV_VAR_SKYBOX_HOR_SPEED

EnvVar_ENV_VAR_SKYBOX_VER_SPEED

EnvVar_ENV_VAR_SKY_COLOR

EnvVar_ENV_VAR_SUN_COLOR

EnvVar_ENV_VAR_WEATHER

EnvVar_ENV_VAR_WEATHER_FADE

EnvVar_ENV_VAR_WEATHER_SPEED

FACE_CONSTS_FACE_COUNT

FACE_CONSTS_FACE_XMAX

FACE_CONSTS_FACE_XMIN

FACE_CONSTS_FACE_YMAX

FACE_CONSTS_FACE_YMIN

FACE_CONSTS_FACE_ZMAX

FACE_CONSTS_FACE_ZMIN

FILENAME_SIZE

FONT_FLAGS_FONT_FLAGS_BOLD

FONT_FLAGS_FONT_FLAGS_NONE

FONT_FLAGS_FONT_FLAGS_PADDING

FONT_FLAGS_FONT_FLAGS_UNDERLINE

File_SeekFrom_FILE_SEEKFROM_BEGIN

File_SeekFrom_FILE_SEEKFROM_CURRENT

File_SeekFrom_FILE_SEEKFROM_END

FogFunc__FOG_EXP

FogFunc__FOG_EXP2

FogFunc__FOG_LINEAR

FpsLimitMethod_FPS_LIMIT_30

FpsLimitMethod_FPS_LIMIT_60

FpsLimitMethod_FPS_LIMIT_120

FpsLimitMethod_FPS_LIMIT_144

FpsLimitMethod_FPS_LIMIT_COUNT

FpsLimitMethod_FPS_LIMIT_NONE

FpsLimitMethod_FPS_LIMIT_VSYNC

GAME_API_VER

GAME_APP_NAME

GAME_APP_TITLE

GAME_APP_VER

GAME_DEF_TICKS

GAME_MAX_CMDARGS

GAME_NET_TICKS

GFX_MAX_INDICES

GFX_MAX_VERTICES

GUI_MAX_SCREENS

GuiAnchor_ANCHOR_CENTRE

GuiAnchor_ANCHOR_CENTRE_MAX

GuiAnchor_ANCHOR_CENTRE_MIN

GuiAnchor_ANCHOR_MAX

GuiAnchor_ANCHOR_MIN

GuiPriority_GUI_PRIORITY_CHAT

GuiPriority_GUI_PRIORITY_DISCONNECT

GuiPriority_GUI_PRIORITY_HUD

GuiPriority_GUI_PRIORITY_INVENTORY

GuiPriority_GUI_PRIORITY_LOADING

GuiPriority_GUI_PRIORITY_MENU

GuiPriority_GUI_PRIORITY_MENUINPUT

GuiPriority_GUI_PRIORITY_OLDLOADING

GuiPriority_GUI_PRIORITY_TABLIST

GuiPriority_GUI_PRIORITY_TEXIDS

GuiPriority_GUI_PRIORITY_TEXPACK

GuiPriority_GUI_PRIORITY_TOUCH

GuiPriority_GUI_PRIORITY_TOUCHMORE

GuiPriority_GUI_PRIORITY_URLWARNING

HALF_CHUNK_SIZE

HOTBAR_MAX_INDEX

HOTKEYS_MAX_COUNT

HotkeyModifiers_HOTKEY_MOD_ALT

HotkeyModifiers_HOTKEY_MOD_CTRL

HotkeyModifiers_HOTKEY_MOD_SHIFT

HttpProgress_HTTP_PROGRESS_FETCHING_DATA

HttpProgress_HTTP_PROGRESS_MAKING_REQUEST

HttpProgress_HTTP_PROGRESS_NOT_WORKING_ON

HttpRequestType_REQUEST_TYPE_GET

HttpRequestType_REQUEST_TYPE_HEAD

HttpRequestType_REQUEST_TYPE_POST

INFLATE_FAST_BITS

INFLATE_MAX_BITS

INFLATE_MAX_CODELENS

INFLATE_MAX_DISTS

INFLATE_MAX_INPUT

INFLATE_MAX_LITS

INFLATE_MAX_LITS_DISTS

INFLATE_WINDOW_MASK

INFLATE_WINDOW_SIZE

INPUTWIDGET_LEN

INPUTWIDGET_MAX_LINES

INPUT_MAX_POINTERS

INVENTORY_BLOCKS_PER_HOTBAR

INVENTORY_HOTBARS

ISOMETRICDRAWER_MAXVERTICES

InputButtons_INPUT_CLIPBOARD_COPY

InputButtons_INPUT_CLIPBOARD_PASTE

InputButtons_INPUT_COUNT

InputButtons_KEY_0

InputButtons_KEY_1

InputButtons_KEY_2

InputButtons_KEY_3

InputButtons_KEY_4

InputButtons_KEY_5

InputButtons_KEY_6

InputButtons_KEY_7

InputButtons_KEY_8

InputButtons_KEY_9

InputButtons_KEY_A

InputButtons_KEY_B

InputButtons_KEY_BACKSLASH

InputButtons_KEY_BACKSPACE

InputButtons_KEY_C

InputButtons_KEY_CAPSLOCK

InputButtons_KEY_COMMA

InputButtons_KEY_D

InputButtons_KEY_DELETE

InputButtons_KEY_DOWN

InputButtons_KEY_E

InputButtons_KEY_END

InputButtons_KEY_ENTER

InputButtons_KEY_EQUALS

InputButtons_KEY_ESCAPE

InputButtons_KEY_F

InputButtons_KEY_F1

InputButtons_KEY_F2

InputButtons_KEY_F3

InputButtons_KEY_F4

InputButtons_KEY_F5

InputButtons_KEY_F6

InputButtons_KEY_F7

InputButtons_KEY_F8

InputButtons_KEY_F9

InputButtons_KEY_F10

InputButtons_KEY_F11

InputButtons_KEY_F12

InputButtons_KEY_F13

InputButtons_KEY_F14

InputButtons_KEY_F15

InputButtons_KEY_F16

InputButtons_KEY_F17

InputButtons_KEY_F18

InputButtons_KEY_F19

InputButtons_KEY_F20

InputButtons_KEY_F21

InputButtons_KEY_F22

InputButtons_KEY_F23

InputButtons_KEY_F24

InputButtons_KEY_G

InputButtons_KEY_H

InputButtons_KEY_HOME

InputButtons_KEY_I

InputButtons_KEY_INSERT

InputButtons_KEY_J

InputButtons_KEY_K

InputButtons_KEY_KP0

InputButtons_KEY_KP1

InputButtons_KEY_KP2

InputButtons_KEY_KP3

InputButtons_KEY_KP4

InputButtons_KEY_KP5

InputButtons_KEY_KP6

InputButtons_KEY_KP7

InputButtons_KEY_KP8

InputButtons_KEY_KP9

InputButtons_KEY_KP_DECIMAL

InputButtons_KEY_KP_DIVIDE

InputButtons_KEY_KP_ENTER

InputButtons_KEY_KP_MINUS

InputButtons_KEY_KP_MULTIPLY

InputButtons_KEY_KP_PLUS

InputButtons_KEY_L

InputButtons_KEY_LALT

InputButtons_KEY_LBRACKET

InputButtons_KEY_LCTRL

InputButtons_KEY_LEFT

InputButtons_KEY_LMOUSE

InputButtons_KEY_LSHIFT

InputButtons_KEY_LWIN

InputButtons_KEY_M

InputButtons_KEY_MENU

InputButtons_KEY_MINUS

InputButtons_KEY_MMOUSE

InputButtons_KEY_N

InputButtons_KEY_NONE

InputButtons_KEY_NUMLOCK

InputButtons_KEY_O

InputButtons_KEY_P

InputButtons_KEY_PAGEDOWN

InputButtons_KEY_PAGEUP

InputButtons_KEY_PAUSE

InputButtons_KEY_PERIOD

InputButtons_KEY_PRINTSCREEN

InputButtons_KEY_Q

InputButtons_KEY_QUOTE

InputButtons_KEY_R

InputButtons_KEY_RALT

InputButtons_KEY_RBRACKET

InputButtons_KEY_RCTRL

InputButtons_KEY_RIGHT

InputButtons_KEY_RMOUSE

InputButtons_KEY_RSHIFT

InputButtons_KEY_RWIN

InputButtons_KEY_S

InputButtons_KEY_SCROLLLOCK

InputButtons_KEY_SEMICOLON

InputButtons_KEY_SLASH

InputButtons_KEY_SPACE

InputButtons_KEY_T

InputButtons_KEY_TAB

InputButtons_KEY_TILDE

InputButtons_KEY_U

InputButtons_KEY_UP

InputButtons_KEY_V

InputButtons_KEY_W

InputButtons_KEY_X

InputButtons_KEY_XBUTTON1

InputButtons_KEY_XBUTTON2

InputButtons_KEY_Y

InputButtons_KEY_Z

Input_Names

Input_TouchMode

Int16_MaxValue

Int32_MaxValue

Int32_MinValue

KeyBind_Defaults

KeyBind__KEYBIND_AUTOROTATE

KeyBind__KEYBIND_AXIS_LINES

KeyBind__KEYBIND_BACK

KeyBind__KEYBIND_BREAK_LIQUIDS

KeyBind__KEYBIND_CHAT

KeyBind__KEYBIND_COUNT

KeyBind__KEYBIND_DELETE_BLOCK

KeyBind__KEYBIND_DROP_BLOCK

KeyBind__KEYBIND_EXT_INPUT

KeyBind__KEYBIND_FLY

KeyBind__KEYBIND_FLY_DOWN

KeyBind__KEYBIND_FLY_UP

KeyBind__KEYBIND_FOG

KeyBind__KEYBIND_FORWARD

KeyBind__KEYBIND_FULLSCREEN

KeyBind__KEYBIND_HALF_SPEED

KeyBind__KEYBIND_HIDE_FPS

KeyBind__KEYBIND_HIDE_GUI

KeyBind__KEYBIND_HOTBAR_SWITCH

KeyBind__KEYBIND_IDOVERLAY

KeyBind__KEYBIND_INVENTORY

KeyBind__KEYBIND_JUMP

KeyBind__KEYBIND_LEFT

KeyBind__KEYBIND_NOCLIP

KeyBind__KEYBIND_PICK_BLOCK

KeyBind__KEYBIND_PLACE_BLOCK

KeyBind__KEYBIND_RESPAWN

KeyBind__KEYBIND_RIGHT

KeyBind__KEYBIND_SCREENSHOT

KeyBind__KEYBIND_SEND_CHAT

KeyBind__KEYBIND_SET_SPAWN

KeyBind__KEYBIND_SMOOTH_CAMERA

KeyBind__KEYBIND_SPEED

KeyBind__KEYBIND_TABLIST

KeyBind__KEYBIND_THIRD_PERSON

KeyBind__KEYBIND_ZOOM_SCROLL

Key_Ascii_Names

Key_Function_Names

KeyboardType_KEYBOARD_TYPE_INTEGER

KeyboardType_KEYBOARD_TYPE_NUMBER

KeyboardType_KEYBOARD_TYPE_PASSWORD

KeyboardType_KEYBOARD_TYPE_TEXT

LOCATIONUPDATE_PITCH

LOCATIONUPDATE_POS

LOCATIONUPDATE_ROTX

LOCATIONUPDATE_ROTZ

LOCATIONUPDATE_YAW

LOPT_PASSWORD

LOPT_SESSION

LOPT_USERNAME

MATH_DEG2RAD

MATH_LARGENUM

MATH_PI

MATH_RAD2DEG

MAX_CUSTOM_MODELS

MAX_CUSTOM_MODEL_ANIMS

MAX_CUSTOM_MODEL_PARTS

MENUINPUTWIDGET_MAX

MILLIS_PER_SEC

MODEL_BOX_VERTICES

MODEL_QUAD_VERTICES

MatrixType__MATRIX_PROJECTION

MatrixType__MATRIX_VIEW

MsgType_MSG_TYPE_ANNOUNCEMENT

MsgType_MSG_TYPE_BIGANNOUNCEMENT

MsgType_MSG_TYPE_BOTTOMRIGHT_1

MsgType_MSG_TYPE_BOTTOMRIGHT_2

MsgType_MSG_TYPE_BOTTOMRIGHT_3

MsgType_MSG_TYPE_CLIENTSTATUS_1

MsgType_MSG_TYPE_CLIENTSTATUS_2

MsgType_MSG_TYPE_NORMAL

MsgType_MSG_TYPE_SMALLANNOUNCEMENT

MsgType_MSG_TYPE_STATUS_1

MsgType_MSG_TYPE_STATUS_2

MsgType_MSG_TYPE_STATUS_3

NATIVE_STR_LEN

NameMode_NAME_MODE_ALL

NameMode_NAME_MODE_ALL_HOVERED

NameMode_NAME_MODE_ALL_UNSCALED

NameMode_NAME_MODE_COUNT

NameMode_NAME_MODE_HOVERED

NameMode_NAME_MODE_NONE

OGG_BUFFER_SIZE

ONSCREEN_BTN_CAMERA

ONSCREEN_BTN_CHAT

ONSCREEN_BTN_DELETE

ONSCREEN_BTN_FLY

ONSCREEN_BTN_HALFSPEED

ONSCREEN_BTN_LIST

ONSCREEN_BTN_NOCLIP

ONSCREEN_BTN_PICK

ONSCREEN_BTN_PLACE

ONSCREEN_BTN_SETSPAWN

ONSCREEN_BTN_SPAWN

ONSCREEN_BTN_SPEED

ONSCREEN_BTN_SWITCH

ONSCREEN_MAX_BTNS

OPCODE__OPCODE_ADD_ENTITY

OPCODE__OPCODE_BULK_BLOCK_UPDATE

OPCODE__OPCODE_COUNT

OPCODE__OPCODE_CUSTOM_BLOCK_LEVEL

OPCODE__OPCODE_DEFINE_BLOCK

OPCODE__OPCODE_DEFINE_BLOCK_EXT

OPCODE__OPCODE_DEFINE_EFFECT

OPCODE__OPCODE_DEFINE_MODEL

OPCODE__OPCODE_DEFINE_MODEL_PART

OPCODE__OPCODE_ENTITY_TELEPORT

OPCODE__OPCODE_ENV_SET_COLOR

OPCODE__OPCODE_ENV_SET_MAP_APPEARANCE

OPCODE__OPCODE_ENV_SET_MAP_PROPERTY

OPCODE__OPCODE_ENV_SET_MAP_URL

OPCODE__OPCODE_ENV_SET_WEATHER

OPCODE__OPCODE_EXT_ADD_ENTITY

OPCODE__OPCODE_EXT_ADD_ENTITY2

OPCODE__OPCODE_EXT_ADD_PLAYER_NAME

OPCODE__OPCODE_EXT_ENTRY

OPCODE__OPCODE_EXT_INFO

OPCODE__OPCODE_EXT_REMOVE_PLAYER_NAME

OPCODE__OPCODE_HACK_CONTROL

OPCODE__OPCODE_HANDSHAKE

OPCODE__OPCODE_HOLD_THIS

OPCODE__OPCODE_KICK

OPCODE__OPCODE_LEVEL_BEGIN

OPCODE__OPCODE_LEVEL_DATA

OPCODE__OPCODE_LEVEL_END

OPCODE__OPCODE_MAKE_SELECTION

OPCODE__OPCODE_MESSAGE

OPCODE__OPCODE_ORI_UPDATE

OPCODE__OPCODE_PING

OPCODE__OPCODE_PLAYER_CLICK

OPCODE__OPCODE_PLUGIN_MESSAGE

OPCODE__OPCODE_RELPOS_AND_ORI_UPDATE

OPCODE__OPCODE_RELPOS_UPDATE

OPCODE__OPCODE_REMOVE_ENTITY

OPCODE__OPCODE_REMOVE_SELECTION

OPCODE__OPCODE_SET_BLOCK

OPCODE__OPCODE_SET_BLOCK_CLIENT

OPCODE__OPCODE_SET_BLOCK_PERMISSION

OPCODE__OPCODE_SET_ENTITY_PROPERTY

OPCODE__OPCODE_SET_HOTBAR

OPCODE__OPCODE_SET_INVENTORY_ORDER

OPCODE__OPCODE_SET_MODEL

OPCODE__OPCODE_SET_PERMISSION

OPCODE__OPCODE_SET_REACH

OPCODE__OPCODE_SET_SPAWNPOINT

OPCODE__OPCODE_SET_TEXT_COLOR

OPCODE__OPCODE_SET_TEXT_HOTKEY

OPCODE__OPCODE_SPAWN_EFFECT

OPCODE__OPCODE_TWO_WAY_PING

OPCODE__OPCODE_UNDEFINE_BLOCK

OPCODE__OPCODE_UNDEFINE_MODEL

OPCODE__OPCODE_VELOCITY_CONTROL

OPT_AUTO_CLOSE_LAUNCHER

OPT_BLACK_TEXT

OPT_BLOCK_PHYSICS

OPT_CAMERA_CLIPPING

OPT_CAMERA_MASS

OPT_CAMERA_SMOOTH

OPT_CHATLINES

OPT_CHAT_LOGGING

OPT_CHAT_SCALE

OPT_CLASSIC_ARM_MODEL

OPT_CLASSIC_CHAT

OPT_CLASSIC_GUI

OPT_CLASSIC_HACKS

OPT_CLASSIC_MODE

OPT_CLASSIC_OPTIONS

OPT_CLASSIC_TABLIST

OPT_CLICKABLE_CHAT

OPT_CPE

OPT_CUSTOM_BLOCKS

OPT_DEFAULT_TEX_PACK

OPT_ENTITY_SHADOW

OPT_FIELD_OF_VIEW

OPT_FONT_NAME

OPT_FORCE_OPENAL

OPT_FPS_LIMIT

OPT_FULL_BLOCK_STEP

OPT_GRAB_CURSOR

OPT_HACKS_ENABLED

OPT_HACK_PERM_MSGS

OPT_HOTBAR_SCALE

OPT_HTTP_ONLY

OPT_INVENTORY_SCALE

OPT_INVERT_MOUSE

OPT_JUMP_VELOCITY

OPT_LANDSCAPE_MODE

OPT_MAX_CHUNK_UPDATES

OPT_MAX_MUSIC_DELAY

OPT_MIN_MUSIC_DELAY

OPT_MIPMAPS

OPT_MODIFIABLE_LIQUIDS

OPT_MUSIC_VOLUME

OPT_NAMES_MODE

OPT_NOCLIP_SLIDE

OPT_PUSHBACK_PLACING

OPT_RAW_INPUT

OPT_RENDER_TYPE

OPT_SENSITIVITY

OPT_SERVER_TEXTURES

OPT_SHOW_BLOCK_IN_HAND

OPT_SHOW_FPS

OPT_SIMPLE_ARMS_ANIM

OPT_SMOOTH_LIGHTING

OPT_SOUND_VOLUME

OPT_SPEED_FACTOR

OPT_TAB_AUTOCOMPLETE

OPT_TOUCH_BUTTONS

OPT_TOUCH_SCALE

OPT_USE_CHAT_FONT

OPT_VIEW_BOBBING

OPT_VIEW_DISTANCE

OPT_WINDOW_HEIGHT

OPT_WINDOW_WIDTH

OPT_WOM_STYLE_HACKS

PACKEDCOL_A_MASK

PACKEDCOL_A_SHIFT

PACKEDCOL_B_MASK

PACKEDCOL_B_SHIFT

PACKEDCOL_G_MASK

PACKEDCOL_G_SHIFT

PACKEDCOL_RGB_MASK

PACKEDCOL_R_MASK

PACKEDCOL_R_SHIFT

PACKEDCOL_SHADE_X

PACKEDCOL_SHADE_YMIN

PACKEDCOL_SHADE_Z

PACKEDCOL_WHITE

PNG_MAX_DIMS

Pointers_Count

REGISTERNEW_URL

RESOURCE_SERVER

RESPAWN_NOT_FOUND

ROPT_IP

ROPT_MPPASS

ROPT_PORT

ROPT_SERVER

ROPT_USER

RotateOrder_ROTATE_ORDER_XYZ

RotateOrder_ROTATE_ORDER_XZY

RotateOrder_ROTATE_ORDER_YZX

RotateOrder_ROTATE_ORDER_ZYX

SECS_PER_DAY

SECS_PER_HOUR

SECS_PER_MIN

SERVICES_SERVER

SIZEOF_VERTEX_COLOURED

SIZEOF_VERTEX_TEXTURED

SKINS_SERVER

SKIN_FETCH_COMPLETED

SKIN_FETCH_DOWNLOADING

SKIN_TYPE_SKIN_64x32

SKIN_TYPE_SKIN_64x64

SKIN_TYPE_SKIN_64x64_SLIM

SKIN_TYPE_SKIN_INVALID

SOPT_SERVICES

STRINGSBUFFER_BUFFER_DEF_SIZE

STRINGSBUFFER_DEF_LEN_MASK

STRINGSBUFFER_DEF_LEN_SHIFT

STRINGSBUFFER_FLAGS_DEF_ELEMS

STRING_INT_CHARS

STRING_SIZE

ShadowMode_SHADOW_MODE_CIRCLE

ShadowMode_SHADOW_MODE_CIRCLE_ALL

ShadowMode_SHADOW_MODE_COUNT

ShadowMode_SHADOW_MODE_NONE

ShadowMode_SHADOW_MODE_SNAP_TO_BLOCK

Socket_PollMode_SOCKET_POLL_READ

Socket_PollMode_SOCKET_POLL_WRITE

SoftKeyboard_SOFT_KEYBOARD_NONE

SoftKeyboard_SOFT_KEYBOARD_RESIZE

SoftKeyboard_SOFT_KEYBOARD_SHIFT

SoundType_SOUND_CLOTH

SoundType_SOUND_COUNT

SoundType_SOUND_GLASS

SoundType_SOUND_GRASS

SoundType_SOUND_GRAVEL

SoundType_SOUND_METAL

SoundType_SOUND_NONE

SoundType_SOUND_SAND

SoundType_SOUND_SNOW

SoundType_SOUND_STONE

SoundType_SOUND_WOOD

TABLIST_MAX_NAMES

TEXPACKS_DIR

TEXTATLAS_MAX_WIDTHS

TEXTGROUPWIDGET_LEN

TEXTGROUPWIDGET_MAX_LINES

TEXTURE_FLAG_DYNAMIC

TEXTURE_FLAG_MANAGED

TEXTWIDGET_MAX

TOUCH_TYPE_ALL

TOUCH_TYPE_BLOCKS

TOUCH_TYPE_CAMERA

TOUCH_TYPE_GUI

TREE_MAX_COUNT

UInt8_MaxValue

UInt16_MaxValue

UNIX_EPOCH

UPDATES_SERVER

UPDATE_FILE

URL_MAX_SIZE

UV2_Scale

VORBIS_MAX_BLOCK_SIZE

VORBIS_MAX_CHANS

VertexFormat__VERTEX_FORMAT_COLOURED

VertexFormat__VERTEX_FORMAT_TEXTURED

WORLD_UUID_LEN

Weather__WEATHER_RAINY

Weather__WEATHER_SNOWY

Weather__WEATHER_SUNNY

WindowState_WINDOW_STATE_FULLSCREEN

WindowState_WINDOW_STATE_MINIMISED

WindowState_WINDOW_STATE_NORMAL

ZIP_MAX_ENTRIES

_NL

controlChars

extendedChars

keybindNames

Statics

Atlas1D

Atlas2D

BlockEvents

Blocks

Camera

ChatEvents

DisplayInfo

Drawer

Drawer2D

Entities

EntityEvents

Env

Game

Gfx

GfxEvents

Gui

InputEvents

Inventory

Models

NetEvents

Physics

PluginMessageEvents

PointerEvents

Protocol

Server

TabList

TabListEvents

TextureEvents

UserEvents

WindowEvents

WindowInfo

World

WorldEvents

Functions

AABB_Make⚠

Bitmap_Scale⚠

Block_FindID⚠

Block_Parse⚠

Block_UNSAFE_GetName⚠

BoxDesc_BuildBox⚠

BoxDesc_BuildRotatedBox⚠

BoxDesc_XQuad⚠

BoxDesc_XQuad2⚠

BoxDesc_YQuad⚠

BoxDesc_YQuad2⚠

BoxDesc_ZQuad⚠

BoxDesc_ZQuad2⚠

CPE_SendPluginMessage⚠

Camera_Register⚠

Chat_Add⚠

Chat_AddOf⚠

Chat_AddRaw⚠

Chat_Send⚠

Clipboard_GetText⚠

Clipboard_SetText⚠

Commands_Register⚠

Convert_CP437ToUnicode

Convert_CodepointToCP437

Convert_ParseBool⚠

Convert_ParseFloat⚠

Convert_ParseInt⚠

Convert_ParseUInt8⚠

Convert_ParseUInt16⚠

Convert_ParseUInt64⚠

DateTime_CurrentLocal⚠

DateTime_CurrentUTC_MS⚠

Deflate_MakeStream⚠

Directory_Create⚠

Directory_Enum⚠

Drawer2D_BmpCopy⚠

Drawer2D_Clear⚠

Drawer2D_DrawText⚠

Drawer2D_MakeFont⚠

Drawer2D_MakeTextTexture⚠

Drawer2D_MakeTexture⚠

Drawer2D_TextHeight⚠

Drawer2D_TextWidth⚠

Drawer_XMax⚠

Drawer_XMin⚠

Drawer_YMax⚠

Drawer_YMin⚠

Drawer_ZMax⚠

Drawer_ZMin⚠

DynamicLib_DescribeError⚠

DynamicLib_Get⚠

DynamicLib_Get2⚠

DynamicLib_Load⚠

DynamicLib_Load2⚠

Entity_GetTransform⚠

Entity_Init⚠

Entity_SetModel⚠

Entity_TouchesAny⚠

Env_Reset⚠

Env_SetCloudsCol⚠

Env_SetCloudsHeight⚠

Env_SetCloudsSpeed⚠

Env_SetEdgeBlock⚠

Env_SetEdgeHeight⚠

Env_SetExpFog⚠

Env_SetFogCol⚠

Env_SetShadowCol⚠

Env_SetSidesBlock⚠

Env_SetSidesOffset⚠

Env_SetSkyCol⚠

Env_SetSkyboxCol⚠

Env_SetSkyboxHorSpeed⚠

Env_SetSkyboxVerSpeed⚠

Env_SetSunCol⚠

Env_SetWeather⚠

Env_SetWeatherFade⚠

Env_SetWeatherSpeed⚠

Event_RaiseBlock⚠

Event_RaiseChat⚠

Event_RaiseEntry⚠

Event_RaiseFloat⚠

Event_RaiseInput⚠

Event_RaiseInt⚠

Event_RaisePluginMessage⚠

Event_RaiseRawMove⚠

Event_RaiseString⚠

Event_RaiseVoid⚠

Event_Register⚠

Event_RegisterBlock⚠

Event_RegisterChat⚠

Event_RegisterEntry⚠

Event_RegisterFloat⚠

Event_RegisterInput⚠

Event_RegisterInt⚠

Event_RegisterPluginMessage⚠

Event_RegisterRawMove⚠

Event_RegisterString⚠

Event_RegisterVoid⚠

Event_Unregister⚠

Event_UnregisterBlock⚠

Event_UnregisterChat⚠

Event_UnregisterEntry⚠

Event_UnregisterFloat⚠

Event_UnregisterInput⚠

Event_UnregisterInt⚠

Event_UnregisterPluginMessage⚠

Event_UnregisterRawMove⚠

Event_UnregisterString⚠

Event_UnregisterVoid⚠

File_Exists⚠

Font_Free⚠

Font_GetNames⚠

GZip_MakeStream⚠

Game_ChangeBlock⚠

Game_UpdateBlock⚠

Gfx_BindIb⚠

Gfx_BindTexture⚠

Gfx_BindVb⚠

Gfx_Clear⚠

Gfx_ClearCol⚠

Gfx_CreateDynamicVb⚠

Gfx_CreateIb⚠

Gfx_CreateTexture⚠

Gfx_CreateVb⚠

Gfx_DeleteIb⚠

Gfx_DeleteTexture⚠

Gfx_DeleteVb⚠

Gfx_DisableMipmaps⚠

Gfx_DisableTextureOffset⚠

Gfx_DrawVb_IndexedTris⚠

Gfx_DrawVb_IndexedTris_Range⚠

Gfx_DrawVb_Lines⚠

Gfx_EnableMipmaps⚠

Gfx_EnableTextureOffset⚠

Gfx_GetFog⚠

Gfx_LoadIdentityMatrix⚠

Gfx_LoadMatrix⚠

Gfx_LockDynamicVb⚠

Gfx_LockVb⚠

Gfx_Make2DQuad⚠

Gfx_SetAlphaArgBlend⚠

Gfx_SetAlphaBlending⚠

Gfx_SetAlphaTest⚠

Gfx_SetColWriteMask⚠

Gfx_SetDepthTest⚠

Gfx_SetDepthWrite⚠

Gfx_SetDynamicVbData⚠

Gfx_SetFaceCulling⚠

Gfx_SetFog⚠

Gfx_SetFogCol⚠

Gfx_SetFogDensity⚠

Gfx_SetFogEnd⚠

Gfx_SetFogMode⚠

Gfx_SetTexturing⚠

Gfx_SetVertexFormat⚠

Gfx_UnlockDynamicVb⚠

Gfx_UnlockVb⚠

Gfx_UpdateDynamicVb_IndexedTris⚠

Gfx_UpdateTexturePart⚠

Gradient_Blend⚠

Gradient_Noise⚠

Gradient_Tint⚠

Gradient_Vertical⚠

Gui_Add⚠

Gui_GetInputGrab⚠

Gui_Remove⚠

IVec3_Floor

IVec3_Max

IVec3_MaxValue

Returns a vector with all components set to Int32_MaxValue.

IVec3_Min

IVec3_ToVec3

Inflate_Init2⚠

Inflate_MakeStream2⚠

Inventory_Get

Gets the block at the nth index in the current hotbar.

Inventory_SelectedBlock

Gets the currently selected block.

Inventory_Set

Sets the block at the nth index in the current hotbar.

KeyBind_IsPressed⚠

LocationUpdate_Clamp

Clamps the given angle so it lies between [0, 360).

LocationUpdate_MakeOri

Makes a location update only containing yaw and pitch.

LocationUpdate_MakePos

Makes a location update only containing position

LocationUpdate_MakePosAndOri

Makes a location update containing position, yaw and pitch.

Map_FindImporter⚠

Map_LoadFrom⚠

Math_Cos⚠

Math_IsPowOf2

Math_NextPowOf2

Math_Sin⚠

Matrix_Mul⚠

Matrix_RotateX⚠

Matrix_RotateY⚠

Matrix_RotateZ⚠

Matrix_Scale⚠

Matrix_Translate⚠

Mem_Alloc⚠

Mem_AllocCleared⚠

Mem_Free⚠

Mem_Realloc⚠

Mem_TryAlloc⚠

Mem_TryAllocCleared⚠

Mem_TryRealloc⚠

ModelPart_Init

Model_ApplyTexture⚠

Model_DrawArmPart⚠

Model_DrawPart⚠

Model_DrawRotate⚠

Model_Get⚠

Model_Init⚠

Model_Register⚠

Model_RegisterTexture⚠

Model_Render⚠

Model_SetupState⚠

Model_UpdateVB⚠

Mutex_Create⚠

Mutex_Free⚠

Mutex_Lock⚠

Mutex_Unlock⚠

NetPlayer_Init⚠

Options_Get⚠

Options_GetBool⚠

Options_GetEnum⚠

Options_GetFloat⚠

Options_GetInt⚠

Options_Reload⚠

Options_SaveIfChanged⚠

Options_Set⚠

Options_SetBool⚠

Options_SetInt⚠

Options_SetString⚠

PackedCol_A_Bits

PackedCol_B_Bits

PackedCol_G_Bits

PackedCol_Lerp⚠

PackedCol_Make

PackedCol_R_Bits

PackedCol_Scale⚠

PackedCol_Tint⚠

Particle_DoRender

Png_Decode⚠

Process_Exit⚠

Process_StartGame⚠

Process_StartOpen⚠

Random_Float⚠

Random_Next⚠

Random_Seed⚠

ScheduledTask_Add⚠

Selections_Add⚠

Selections_Remove⚠

Socket_Available⚠

Socket_Close⚠

Socket_Connect⚠

Socket_GetError⚠

Socket_Poll⚠

Socket_Read⚠

Socket_ValidAddress⚠

Socket_Write⚠

Stopwatch_ElapsedMicroseconds⚠

Stopwatch_Measure⚠

Stream_CreateFile⚠

Stream_FromFile⚠

Stream_OpenFile⚠

Stream_Read⚠

Stream_ReadLine⚠

Stream_ReadonlyBuffered⚠

Stream_ReadonlyMemory⚠

Stream_ReadonlyPortion⚠

Stream_Write⚠

Stream_WriteLine⚠

String_Append⚠

String_AppendBool⚠

String_AppendColorless⚠

String_AppendConst⚠

String_AppendFloat⚠

String_AppendHex⚠

String_AppendInt⚠

String_AppendPaddedInt⚠

String_AppendString⚠

String_AppendUInt32⚠

String_CalcLen⚠

String_CaselessContains⚠

String_CaselessEnds⚠

String_CaselessEquals⚠

String_CaselessEqualsConst⚠

String_CaselessStarts⚠

String_Compare⚠

String_Copy⚠

String_CopyToRaw⚠

String_DeleteAt⚠

String_Equals⚠

String_Format1⚠

String_Format2⚠

String_Format3⚠

String_Format4⚠

String_FromReadonly⚠

String_IndexOfAt⚠

String_IndexOfConst⚠

String_Init⚠

String_InsertAt⚠

String_LastIndexOfAt⚠

String_UNSAFE_Separate⚠

String_UNSAFE_Split⚠

String_UNSAFE_SplitBy⚠

String_UNSAFE_Substring⚠

String_UNSAFE_SubstringAt⚠

String_UNSAFE_TrimEnd⚠

String_UNSAFE_TrimStart⚠

StringsBuffer_Add⚠

StringsBuffer_Remove⚠

StringsBuffer_UNSAFE_Get⚠

TabList_Remove⚠

TabList_Set⚠

TexturePack_Extract⚠

Thread_Detach⚠

Thread_Join⚠

Thread_Sleep⚠

Thread_Start⚠

UNSAFE_GetString⚠

Vec3_BigPos

Vec3_Create3

Vec3_Equals

Whether all of the components of the two vectors are equal.

Vec3_GetDirVector

Returns a normalised vector facing in the direction described by the given yaw and pitch.

Vec3_LengthSquared

Returns the squared length of the vector. Squared length can be used for comparison, to avoid a costly sqrt() However, you must sqrt() this when adding lengths.

Vec3_Lerp

Linearly interpolates components of two vectors.

Vec3_Normalize

Scales all components of a vector to lie in [-1, 1]

Vec3_RotateX

Vec3_RotateY

Vec3_RotateY3

Vec3_RotateZ

Vec3_Transform

Transforms a vector by the given matrix.

Vec3_TransformY

Same as Vec3_Transform, but faster since X and Z are assumed as 0.

Waitable_Create⚠

Waitable_Free⚠

Waitable_Signal⚠

Waitable_Wait⚠

Waitable_WaitFor⚠

Window_SetTitle⚠

Window_ShowDialog⚠

World_Blocks

World_Blocks2

World_GetBlock

World_NewMap⚠

World_Pack

World_SetNewMap⚠

ZLib_MakeStream⚠

Zip_Extract⚠

Zip_Init⚠

Type Definitions

BLOCKID

BitmapCol

BlockID

BlockRaw

Button_Get

Button_Set

CC_ERRORS

CollideType

CustomModelAnimAxis

CustomModelAnimType

Directory_EnumCallback

DrawType

EntityID

EntityType

Entity_TouchesCondition

EntryList_Filter

EnvVar

Event_Block_Callback

Event_Chat_Callback

Event_Entry_Callback

Event_Float_Callback

Event_Input_Callback

Event_Int_Callback

Event_PluginMessage_Callback

Event_RawMove_Callback

Event_String_Callback

Event_Void_Callback

FACE_CONSTS

FONT_FLAGS

Face

File_SeekFrom

FogFunc

FogFunc_

FpsLimitMethod

GfxResourceID

GuiAnchor

GuiPriority

HotkeyModifiers

HttpProgress

HttpRequestType

IMapImporter

IVec3

InputButtons

JsonOnNew

JsonOnValue

KeyBind

KeyBind_

KeyboardType

LScreen_Func

LWidget_Func

Logger_DescribeError

Logger_DoWarn

MatrixType

MatrixType_

MsgType

NameMode

Net_Handler

OPCODE_

OpenFileDialogCallback

PackedCol

PhysicsHandler

Png_RowGetter

RNGState

Rect2D

RotateOrder

SKIN_TYPE

ScheduledTaskCallback

ShadowMode

Socket_PollMode

SoftKeyboard

SoundType

SpecialInputAppendFunc

TextGroupWidget_Get

TextureLoc

TextureRec

Thread_StartFunc

TimeMS

Vec2

Vec3

VertexFormat

VertexFormat_

Weather_

Widget_LeftClick

WindowState

cc_bool

cc_codepoint

cc_file

cc_int8

cc_int16

cc_int32

cc_int64

cc_result

cc_socket

cc_string

cc_uint8

cc_uint16

cc_uint32

cc_uint64

cc_uintptr

cc_unichar

Unions

InflateState__bindgen_ty_1

IntAndFloat

MenuInputDesc__bindgen_ty_1

Stream__bindgen_ty_1