Crate classicube_sys

classicube-sys

Rust bindings to ClassiCube

Documentation

Example

Add this to Cargo.toml:

[lib]
crate-type = ["cdylib"]

[dependencies]
classicube-sys = { git = "https://github.com/SpiralP/rust-classicube-sys.git" }

lib.rs:

use classicube_sys::*;
use std::{os::raw::c_int, ptr};

extern "C" fn init() {
    let owned_string = OwnedString::new("hello from rust!");
    unsafe {
        Chat_Add(owned_string.as_cc_string());
    }
}

#[unsafe(no_mangle)]
pub static Plugin_ApiVersion: c_int = 1;

#[unsafe(no_mangle)]
pub static mut Plugin_Component: IGameComponent = IGameComponent {
    // Called when the game is being loaded.
    Init: Some(init),
    // Called when the component is being freed. (e.g. due to game being closed)
    Free: None,
    // Called to reset the component's state. (e.g. reconnecting to server)
    Reset: None,
    // Called to update the component's state when the user begins loading a new map.
    OnNewMap: None,
    // Called to update the component's state when the user has finished loading a new map.
    OnNewMapLoaded: None,
    // Next component in linked list of components.
    next: ptr::null_mut(),
};

Copy the .dll/.so/.dylib from target/{debug,release}/ to the plugins folder where your ClassiCube executable is found

References

Example library usage

ClassiCube source

ClassiCube Plugin help

ClassiCube Plugin examples

Re-exports

pub use crate::screen::OwnedScreen;

Modules

screen

Macros

BoxDesc_Bounds

BoxDesc_Box

gives ((x, y, z), (x1, y1, z1, x2, y2, z2))

BoxDesc_Dim

BoxDesc_Dims

gives (x, y, z)

BoxDesc_Rot

gives (x, y, z)

BoxDesc_Tex

Model_RetAABB

Model_RetSize

Vec3_AddBy

Vec3_IsZero

Whether all components of a 3D vector are 0

Vec3_Mul1By

Vec3_Mul3By

Vec3_Set

Sets the x, y, and z components of a 3D vector

Vec3_SubBy

Structs

AABB

AnimatedComp

AudioChunk

AudioContext

AudioData

BindMapping_

Bitmap

BoxDesc

ButtonWidget

Camera

ChatCommand

ChatInputWidget

CheckUpdateData

ChunkInfo

ChunkPartInfo

Codebook

CollisionsComp

Context2D

CustomModel

CustomModelAnim

CustomModelPart

CustomModelPartDef

CustomParticleEffect

DateTime

DeflateState

DrawTextArgs

DynamicLibSym

Entity

EntityLocation

EntityVTABLE

Event_Block

Event_Chat

Event_Entry

Event_Float

Event_Input

Event_Int

Event_LightingMode

Event_PadAxis

Event_PluginMessage

Event_RawMove

Event_String

Event_Void

FetchFlagsData

FetchServerData

FetchServersData

FetchUpdateData

Flag

Floor

FontDesc

GZipHeader

GZipState

GameVersion

GamepadDevice

GetTokenTaskData

HacksComp

HotbarWidget

HotkeyData

HttpRequest

HuffmanTable

IGameComponent

IVec3_

InflateState

InputDevice

InputWidget

InterpComp

JsonContext

KBLayout

LButton

LCheckbox

LInput

LLabel

LLayout

LLine

LScreen

LSlider

LTable

LTableCell

LTableColumn

LWebTask

LWidget

LWidgetVTABLE

LauncherTheme

LocalPlayer

LocalPlayerInput

LocationUpdate

MapGenerator

MapImporter

Mapping

Matrix

MenuInputDesc

MenuInputDesc__bindgen_ty_1__bindgen_ty_1

MenuInputDesc__bindgen_ty_1__bindgen_ty_2

MenuInputDesc__bindgen_ty_1__bindgen_ty_3

MenuInputDesc__bindgen_ty_1__bindgen_ty_4

MenuInputVTABLE

MenuOptionsScreen

Mode

Model

ModelPart

ModelTex

ModelVertex

NetInterpAngles

NetInterpComp

NetPlayer

OggState

OpenFileDialogArgs

OpenKeyboardArgs

OwnedBitmap

OwnedChatCommand

OwnedContext2D

OwnedGfxTexture

OwnedGfxVertexBuffer

OwnedString

OwnedTexture

Particle

PhysicsComp

Physics_

Pointer

Queue

RayTracer

Rect2D_

Residue

ResumeInfo

SaveFileDialogArgs

ScheduledTask

Screen

ScreenVTABLE

ScrollbarWidget

SearcherState

ServerInfo

SignInTaskData

SimpleButtonDesc

SpecialInputTab

SpecialInputWidget

Stream

Stream__bindgen_ty_1__bindgen_ty_1

Stream__bindgen_ty_1__bindgen_ty_2

Stream__bindgen_ty_1__bindgen_ty_3

Stream__bindgen_ty_1__bindgen_ty_4

StringsBuffer

TableWidget

TextAtlas

TextGroupWidget

TextInputWidget

TextWidget

Texture

TextureEntry

TextureRec_

TiltComp

UpdaterBuild

UpdaterInfo

Vec4

Vec2_

Vec3_

VertexColoured

VertexTextured

VorbisState

VorbisWindow

Widget

WidgetVTABLE

ZLibHeader

ZLibState

ZipEntry

_Atlas1DData

_Atlas2DData

_BlockEventsList

_BlockLists

_CameraData

_ChatEventsList

_ControllerEventsList

_DisplayData

_Drawer2DData

_DrawerData

_EntitiesData

_EntityEventsList

_EnvData

_GameData

_GfxData

_GfxEventsList

_GuiData

_InputEventsList

_InputState

_InventoryData

_Lighting

_ModelsData

_NetEventsList

_PointerEventsList

_ProtocolData

_ServerConnectionData

_TabListData

_TabListEventsList

_TextureEventsList

_UserEventsList

_WindowEventsList

_WorldData

_WorldEventsList

__BindgenBitfieldUnit

cc_sockaddr_

cc_string_

cc_window

cc_winstring_

imdct_state

Constants

ATLAS1D_MAX_ATLASES

ATLAS2D_MASK

ATLAS2D_MAX_ROWS_COUNT

ATLAS2D_SHIFT

ATLAS2D_TILES_PER_ROW

AUDIO_MAX_BUFFERS

AXIS_BEHAVIOUR_AXIS_BEHAVIOUR_CAMERA

AXIS_BEHAVIOUR_AXIS_BEHAVIOUR_MOVEMENT

AXIS_SENSITIVITY_AXIS_SENSI_HIGH

AXIS_SENSITIVITY_AXIS_SENSI_HIGHER

AXIS_SENSITIVITY_AXIS_SENSI_LOW

AXIS_SENSITIVITY_AXIS_SENSI_LOWER

AXIS_SENSITIVITY_AXIS_SENSI_NORMAL

BITMAPCOLOR_A_MASK

BITMAPCOLOR_A_SHIFT

BITMAPCOLOR_B_MASK

BITMAPCOLOR_B_SHIFT

BITMAPCOLOR_G_MASK

BITMAPCOLOR_G_SHIFT

BITMAPCOLOR_RGB_MASK

BITMAPCOLOR_R_MASK

BITMAPCOLOR_R_SHIFT

BITMAPCOLOR_SIZE

BLOCKID_BLOCK_AIR

BLOCKID_BLOCK_AQUA

BLOCKID_BLOCK_BEDROCK

BLOCKID_BLOCK_BLACK

BLOCKID_BLOCK_BLUE

BLOCKID_BLOCK_BOOKSHELF

BLOCKID_BLOCK_BRICK

BLOCKID_BLOCK_BROWN

BLOCKID_BLOCK_BROWN_SHROOM

BLOCKID_BLOCK_CERAMIC_TILE

BLOCKID_BLOCK_COAL_ORE

BLOCKID_BLOCK_COBBLE

BLOCKID_BLOCK_COBBLE_SLAB

BLOCKID_BLOCK_COUNT

BLOCKID_BLOCK_CRATE

BLOCKID_BLOCK_CYAN

BLOCKID_BLOCK_DANDELION

BLOCKID_BLOCK_DEEP_BLUE

BLOCKID_BLOCK_DIRT

BLOCKID_BLOCK_DOUBLE_SLAB

BLOCKID_BLOCK_FIRE

BLOCKID_BLOCK_FOREST_GREEN

BLOCKID_BLOCK_GLASS

BLOCKID_BLOCK_GOLD

BLOCKID_BLOCK_GOLD_ORE

BLOCKID_BLOCK_GRASS

BLOCKID_BLOCK_GRAVEL

BLOCKID_BLOCK_GRAY

BLOCKID_BLOCK_GREEN

BLOCKID_BLOCK_ICE

BLOCKID_BLOCK_INDIGO

BLOCKID_BLOCK_IRON

BLOCKID_BLOCK_IRON_ORE

BLOCKID_BLOCK_LAVA

BLOCKID_BLOCK_LEAVES

BLOCKID_BLOCK_LIGHT_PINK

BLOCKID_BLOCK_LIME

BLOCKID_BLOCK_LOG

BLOCKID_BLOCK_MAGENTA

BLOCKID_BLOCK_MAGMA

BLOCKID_BLOCK_MAX_CPE

BLOCKID_BLOCK_MAX_DEFINED

BLOCKID_BLOCK_MAX_ORIGINAL

BLOCKID_BLOCK_MOSSY_ROCKS

BLOCKID_BLOCK_OBSIDIAN

BLOCKID_BLOCK_ORANGE

BLOCKID_BLOCK_PILLAR

BLOCKID_BLOCK_PINK

BLOCKID_BLOCK_RED

BLOCKID_BLOCK_RED_SHROOM

BLOCKID_BLOCK_ROPE

BLOCKID_BLOCK_ROSE

BLOCKID_BLOCK_SAND

BLOCKID_BLOCK_SANDSTONE

BLOCKID_BLOCK_SAPLING

BLOCKID_BLOCK_SLAB

BLOCKID_BLOCK_SNOW

BLOCKID_BLOCK_SPONGE

BLOCKID_BLOCK_STILL_LAVA

BLOCKID_BLOCK_STILL_WATER

BLOCKID_BLOCK_STONE

BLOCKID_BLOCK_STONE_BRICK

BLOCKID_BLOCK_TEAL

BLOCKID_BLOCK_TNT

BLOCKID_BLOCK_TURQUOISE

BLOCKID_BLOCK_VIOLET

BLOCKID_BLOCK_WATER

BLOCKID_BLOCK_WHITE

BLOCKID_BLOCK_WOOD

BLOCKID_BLOCK_YELLOW

BUTTONWIDGET_MAX

CC_AUD_BACKEND

CC_AUD_BACKEND_OPENAL

CC_AUD_BACKEND_OPENSLES

CC_AUD_BACKEND_WINMM

CC_ERRORS_AL_ERR_INIT_CONTEXT

CC_ERRORS_AL_ERR_INIT_DEVICE

CC_ERRORS_AUDIO_ERR_MP3_SIG

CC_ERRORS_CW_ERR_ROOT_TAG

CC_ERRORS_CW_ERR_STRING_LEN

CC_ERRORS_CW_ERR_UUID_LEN

CC_ERRORS_DAT_ERR_IDENTIFIER

CC_ERRORS_DAT_ERR_JIDENTIFIER

CC_ERRORS_DAT_ERR_JVERSION

CC_ERRORS_DAT_ERR_ROOT_OBJECT

CC_ERRORS_DAT_ERR_VERSION

CC_ERRORS_ERROR_BASE

CC_ERRORS_ERR_DOWNLOAD_INVALID

CC_ERRORS_ERR_END_OF_STREAM

CC_ERRORS_ERR_INVALID_ARGUMENT

CC_ERRORS_ERR_INVALID_DATA_URL

CC_ERRORS_ERR_INVALID_OPEN_URL

CC_ERRORS_ERR_NOT_SUPPORTED

CC_ERRORS_ERR_NO_AUDIO_OUTPUT

CC_ERRORS_ERR_NO_NETWORKING

CC_ERRORS_ERR_OUT_OF_MEMORY

CC_ERRORS_FCM_ERR_IDENTIFIER

CC_ERRORS_FCM_ERR_REVISION

CC_ERRORS_GZIP_ERR_FLAGS

CC_ERRORS_GZIP_ERR_HEADER1

CC_ERRORS_GZIP_ERR_HEADER2

CC_ERRORS_GZIP_ERR_METHOD

CC_ERRORS_HTTP_ERR_CHUNK_SIZE

CC_ERRORS_HTTP_ERR_INVALID_BODY

CC_ERRORS_HTTP_ERR_NO_RESPONSE

CC_ERRORS_HTTP_ERR_NO_SSL

CC_ERRORS_HTTP_ERR_REDIRECTS

CC_ERRORS_HTTP_ERR_RELATIVE

CC_ERRORS_HTTP_ERR_TRUNCATED

CC_ERRORS_INF_ERR_BLOCKTYPE

CC_ERRORS_INF_ERR_INVALID_CODE

CC_ERRORS_INF_ERR_LEN_VERIFY

CC_ERRORS_INF_ERR_NUM_CODES

CC_ERRORS_INF_ERR_REPEAT_BEG

CC_ERRORS_INF_ERR_REPEAT_END

CC_ERRORS_JAVA_ERR_INVALID_TYPECODE

CC_ERRORS_JAVA_ERR_JCLASSES_COUNT

CC_ERRORS_JAVA_ERR_JCLASS_ANNOTATION

CC_ERRORS_JAVA_ERR_JCLASS_FIELDS

CC_ERRORS_JAVA_ERR_JCLASS_REFERENCE

CC_ERRORS_JAVA_ERR_JCLASS_TYPE

CC_ERRORS_JAVA_ERR_JFIELD_CLASS_NAME

CC_ERRORS_JAVA_ERR_JOBJECT_FLAGS

CC_ERRORS_JAVA_ERR_JSTRING_LEN

CC_ERRORS_JAVA_ERR_JVALUE_TYPE

CC_ERRORS_LVL_ERR_VERSION

CC_ERRORS_NBT_ERR_ARR_TOO_SMALL

CC_ERRORS_NBT_ERR_EXPECTED_ARR

CC_ERRORS_NBT_ERR_EXPECTED_F32

CC_ERRORS_NBT_ERR_EXPECTED_I8

CC_ERRORS_NBT_ERR_EXPECTED_I16

CC_ERRORS_NBT_ERR_EXPECTED_I32

CC_ERRORS_NBT_ERR_EXPECTED_STR

CC_ERRORS_NBT_ERR_UNKNOWN

CC_ERRORS_OGG_ERR_INVALID_SIG

CC_ERRORS_OGG_ERR_VERSION

CC_ERRORS_PNG_ERR_16BITSAMPLES

CC_ERRORS_PNG_ERR_COMP_METHOD

CC_ERRORS_PNG_ERR_FILTER

CC_ERRORS_PNG_ERR_INTERLACED

CC_ERRORS_PNG_ERR_INVALID_COL_BPP

CC_ERRORS_PNG_ERR_INVALID_HDR_SIZE

CC_ERRORS_PNG_ERR_INVALID_SCANLINE

CC_ERRORS_PNG_ERR_INVALID_SIG

CC_ERRORS_PNG_ERR_NO_DATA

CC_ERRORS_PNG_ERR_PAL_SIZE

CC_ERRORS_PNG_ERR_REACHED_IEND

CC_ERRORS_PNG_ERR_TOO_TALL

CC_ERRORS_PNG_ERR_TOO_WIDE

CC_ERRORS_PNG_ERR_TRANS_COUNT

CC_ERRORS_PNG_ERR_TRANS_INVALID

CC_ERRORS_SFD_ERR_NEED_DEFAULT_NAME

CC_ERRORS_SOCK_ERR_UNKNOWN_HOST

CC_ERRORS_SSL_ERR_CONTEXT_DEAD

CC_ERRORS_VORBIS_ERR_BLOCKSIZE

CC_ERRORS_VORBIS_ERR_CHANS

CC_ERRORS_VORBIS_ERR_CODEBOOK_ENTRY

CC_ERRORS_VORBIS_ERR_CODEBOOK_LOOKUP

CC_ERRORS_VORBIS_ERR_CODEBOOK_SYNC

CC_ERRORS_VORBIS_ERR_FLOOR_TYPE

CC_ERRORS_VORBIS_ERR_FRAME_TYPE

CC_ERRORS_VORBIS_ERR_FRAMING

CC_ERRORS_VORBIS_ERR_MAPPING_CHANS

CC_ERRORS_VORBIS_ERR_MAPPING_RESERVED

CC_ERRORS_VORBIS_ERR_MAPPING_TYPE

CC_ERRORS_VORBIS_ERR_MODE_TRANSFORM

CC_ERRORS_VORBIS_ERR_MODE_TYPE

CC_ERRORS_VORBIS_ERR_MODE_WINDOW

CC_ERRORS_VORBIS_ERR_RESIDUE_TYPE

CC_ERRORS_VORBIS_ERR_TIME_TYPE

CC_ERRORS_VORBIS_ERR_VERSION

CC_ERRORS_VORBIS_ERR_WRONG_HEADER

CC_ERRORS_WAV_ERR_DATA_TYPE

CC_ERRORS_WAV_ERR_SAMPLE_BITS

CC_ERRORS_WAV_ERR_STREAM_HDR

CC_ERRORS_WAV_ERR_STREAM_TYPE

CC_ERRORS_ZIP_ERR_FILENAME_LEN

CC_ERRORS_ZIP_ERR_INVALID_CENTRAL_DIR

CC_ERRORS_ZIP_ERR_INVALID_LOCAL_DIR

CC_ERRORS_ZIP_ERR_NO_END_OF_CENTRAL_DIR

CC_ERRORS_ZIP_ERR_SEEK_CENTRAL_DIR

CC_ERRORS_ZIP_ERR_SEEK_END_OF_CENTRAL_DIR

CC_ERRORS_ZIP_ERR_SEEK_LOCAL_DIR

CC_ERRORS_ZIP_ERR_TOO_MANY_ENTRIES

CC_ERRORS_ZLIB_ERR_FLAGS

CC_ERRORS_ZLIB_ERR_METHOD

CC_GFX_BACKEND

CC_GFX_BACKEND_D3D9

CC_GFX_BACKEND_D3D11

CC_GFX_BACKEND_GL1

CC_GFX_BACKEND_GL2

CC_GFX_BACKEND_SOFTGPU

CC_GFX_BACKEND_VULKAN

CC_NET_BACKEND

CC_NET_BACKEND_BUILTIN

CC_NET_BACKEND_LIBCURL

CC_SOCKETADDR_MAXSIZE

CC_SSL_BACKEND

CC_SSL_BACKEND_BEARSSL

CC_SSL_BACKEND_NONE

CC_SSL_BACKEND_SCHANNEL

CC_WIN_BACKEND

CC_WIN_BACKEND_ANDROID

CC_WIN_BACKEND_BEOS

CC_WIN_BACKEND_COCOA

CC_WIN_BACKEND_SDL2

CC_WIN_BACKEND_SDL3

CC_WIN_BACKEND_TERMINAL

CC_WIN_BACKEND_WIN32

CC_WIN_BACKEND_X11

CHATLOG_TIME_MASK

CHUNK_MASK

CHUNK_MAX

CHUNK_SHIFT

CHUNK_SIZE

CHUNK_SIZE_2

CHUNK_SIZE_3

COMMAND_FLAG_SINGLEPLAYER_ONLY

COMMAND_FLAG_UNSPLIT_ARGS

CollideType_COLLIDE_CLIMB

CollideType_COLLIDE_ICE

CollideType_COLLIDE_LAVA

CollideType_COLLIDE_LIQUID

CollideType_COLLIDE_NONE

CollideType_COLLIDE_SLIPPERY_ICE

CollideType_COLLIDE_SOLID

CollideType_COLLIDE_WATER

CustomModelAnimAxis_CustomModelAnimAxis_X

CustomModelAnimAxis_CustomModelAnimAxis_Y

CustomModelAnimAxis_CustomModelAnimAxis_Z

CustomModelAnimType_CustomModelAnimType_FlipRotate

CustomModelAnimType_CustomModelAnimType_FlipRotateVelocity

CustomModelAnimType_CustomModelAnimType_FlipSize

CustomModelAnimType_CustomModelAnimType_FlipSizeVelocity

CustomModelAnimType_CustomModelAnimType_FlipTranslate

CustomModelAnimType_CustomModelAnimType_FlipTranslateVelocity

CustomModelAnimType_CustomModelAnimType_Head

CustomModelAnimType_CustomModelAnimType_LeftArmX

CustomModelAnimType_CustomModelAnimType_LeftArmZ

CustomModelAnimType_CustomModelAnimType_LeftLegX

CustomModelAnimType_CustomModelAnimType_None

CustomModelAnimType_CustomModelAnimType_RightArmX

CustomModelAnimType_CustomModelAnimType_RightArmZ

CustomModelAnimType_CustomModelAnimType_RightLegX

CustomModelAnimType_CustomModelAnimType_SinRotate

CustomModelAnimType_CustomModelAnimType_SinRotateVelocity

CustomModelAnimType_CustomModelAnimType_SinSize

CustomModelAnimType_CustomModelAnimType_SinSizeVelocity

CustomModelAnimType_CustomModelAnimType_SinTranslate

CustomModelAnimType_CustomModelAnimType_SinTranslateVelocity

CustomModelAnimType_CustomModelAnimType_Spin

CustomModelAnimType_CustomModelAnimType_SpinVelocity

DEFAULT_AUD_BACKEND

DEFAULT_GFX_BACKEND

DEFAULT_MAX_VIEWDIST

DEFAULT_MUSIC_VOLUME

DEFAULT_NET_BACKEND

DEFAULT_RESUME_ARG

DEFAULT_SINGLEPLAYER_ARG

DEFAULT_SOUNDS_VOLUME

DEFAULT_SSL_BACKEND

DEFAULT_UI_SCALE_X

DEFAULT_UI_SCALE_Y

DEFAULT_USERNAME

DEFAULT_VIEWDIST

DEFAULT_WIN_BACKEND

DEFLATE_BLOCK_SIZE

DEFLATE_BUFFER_SIZE

DEFLATE_HASH_MASK

DEFLATE_HASH_SIZE

DEFLATE_OUT_SIZE

DRAWER2D_MAX_COLORS

DRAWER2D_MAX_TEXT_LENGTH

DrawType_DRAW_GAS

DrawType_DRAW_OPAQUE

DrawType_DRAW_SPRITE

DrawType_DRAW_TRANSLUCENT

DrawType_DRAW_TRANSPARENT

DrawType_DRAW_TRANSPARENT_THICK

ENTITIES_MAX_COUNT

ENTITIES_SELF_ID

ENTITY_ADJUSTMENT

ENTITY_FLAG_CLASSIC_ADJUST

ENTITY_FLAG_HAS_MODELVB

ENTITY_FLAG_MODEL_RESTRICTED_SCALE

ENV_LEGACY

ENV_MINIMAL

EVENT_MAX_CALLBACKS

EXTCHUNK_SIZE

EXTCHUNK_SIZE_2

EXTCHUNK_SIZE_3

EntityType_ENTITY_TYPE_NONE

EntityType_ENTITY_TYPE_PLAYER

EnvVar_ENV_VAR_CLOUDS_COLOR

EnvVar_ENV_VAR_CLOUDS_HEIGHT

EnvVar_ENV_VAR_CLOUDS_SPEED

EnvVar_ENV_VAR_EDGE_BLOCK

EnvVar_ENV_VAR_EDGE_HEIGHT

EnvVar_ENV_VAR_EXP_FOG

EnvVar_ENV_VAR_FOG_COLOR

EnvVar_ENV_VAR_LAMPLIGHT_COLOR

EnvVar_ENV_VAR_LAVALIGHT_COLOR

EnvVar_ENV_VAR_SHADOW_COLOR

EnvVar_ENV_VAR_SIDES_BLOCK

EnvVar_ENV_VAR_SIDES_OFFSET

EnvVar_ENV_VAR_SKYBOX_COLOR

EnvVar_ENV_VAR_SKYBOX_HOR_SPEED

EnvVar_ENV_VAR_SKYBOX_VER_SPEED

EnvVar_ENV_VAR_SKY_COLOR

EnvVar_ENV_VAR_SUN_COLOR

EnvVar_ENV_VAR_WEATHER

EnvVar_ENV_VAR_WEATHER_FADE

EnvVar_ENV_VAR_WEATHER_SPEED

FACE_CONSTS_FACE_BIT_XMAX

FACE_CONSTS_FACE_BIT_XMIN

FACE_CONSTS_FACE_BIT_YMAX

FACE_CONSTS_FACE_BIT_YMIN

FACE_CONSTS_FACE_BIT_ZMAX

FACE_CONSTS_FACE_BIT_ZMIN

FACE_CONSTS_FACE_COUNT

FACE_CONSTS_FACE_XMAX

FACE_CONSTS_FACE_XMIN

FACE_CONSTS_FACE_YMAX

FACE_CONSTS_FACE_YMIN

FACE_CONSTS_FACE_ZMAX

FACE_CONSTS_FACE_ZMIN

FANCY_AO

FANCY_LIGHTING_LAMP_MASK

FANCY_LIGHTING_LAMP_SHIFT

FANCY_LIGHTING_LEVELS

FANCY_LIGHTING_MAX_LEVEL

FILENAME_SIZE

FONT_FLAGS_FONT_FLAGS_BOLD

FONT_FLAGS_FONT_FLAGS_NONE

FONT_FLAGS_FONT_FLAGS_PADDING

FONT_FLAGS_FONT_FLAGS_UNDERLINE

File_SeekFrom_FILE_SEEKFROM_BEGIN

File_SeekFrom_FILE_SEEKFROM_CURRENT

File_SeekFrom_FILE_SEEKFROM_END

FogFunc__FOG_EXP

FogFunc__FOG_EXP2

FogFunc__FOG_LINEAR

FpsLimitMethod_FPS_LIMIT_30

FpsLimitMethod_FPS_LIMIT_60

FpsLimitMethod_FPS_LIMIT_120

FpsLimitMethod_FPS_LIMIT_144

FpsLimitMethod_FPS_LIMIT_COUNT

FpsLimitMethod_FPS_LIMIT_NONE

FpsLimitMethod_FPS_LIMIT_VSYNC

GAME_API_VER

GAME_APP_NAME

GAME_APP_TITLE

GAME_APP_VER

GAME_DEF_TICKS

GAME_MAX_CMDARGS

GAME_NET_TICKS

GAME_VERSION__VERSION_0017

GAME_VERSION__VERSION_0019

GAME_VERSION__VERSION_0023

GAME_VERSION__VERSION_0030

GAME_VERSION__VERSION_CPE

GFX_MAX_INDICES

GFX_MAX_VERTICES

GUI_MAX_CHATLINES

GUI_MAX_SCREENS

Game_NumStates

GfxBuffers__GFX_BUFFER_COLOR

GfxBuffers__GFX_BUFFER_DEPTH

GuiAnchor_ANCHOR_CENTRE

GuiAnchor_ANCHOR_CENTRE_MAX

GuiAnchor_ANCHOR_CENTRE_MIN

GuiAnchor_ANCHOR_MAX

GuiAnchor_ANCHOR_MIN

GuiPriority_GUI_PRIORITY_CHAT

GuiPriority_GUI_PRIORITY_DISCONNECT

GuiPriority_GUI_PRIORITY_HUD

GuiPriority_GUI_PRIORITY_INVENTORY

GuiPriority_GUI_PRIORITY_LOADING

GuiPriority_GUI_PRIORITY_MENU

GuiPriority_GUI_PRIORITY_MENUINPUT

GuiPriority_GUI_PRIORITY_OLDLOADING

GuiPriority_GUI_PRIORITY_TABLIST

GuiPriority_GUI_PRIORITY_TEXIDS

GuiPriority_GUI_PRIORITY_TEXPACK

GuiPriority_GUI_PRIORITY_TOUCH

GuiPriority_GUI_PRIORITY_TOUCHMORE

GuiPriority_GUI_PRIORITY_URLWARNING

Gui_TouchUI

HALF_CHUNK_SIZE

HOTBAR_CORE_VERTICES

HOTBAR_MAX_INDEX

HOTBAR_MAX_VERTICES

HOTKEYS_MAX_COUNT

HOTKEY_FLAG_AUTO_DEFINED

HOTKEY_FLAG_STAYS_OPEN

HTTP_FLAG_NOCACHE

HTTP_FLAG_PRIORITY

HotkeyModifiers_HOTKEY_MOD_ALT

HotkeyModifiers_HOTKEY_MOD_CTRL

HotkeyModifiers_HOTKEY_MOD_SHIFT

HttpProgress_HTTP_PROGRESS_FETCHING_DATA

HttpProgress_HTTP_PROGRESS_MAKING_REQUEST

HttpProgress_HTTP_PROGRESS_NOT_WORKING_ON

HttpRequestType_REQUEST_TYPE_GET

HttpRequestType_REQUEST_TYPE_HEAD

HttpRequestType_REQUEST_TYPE_POST

INFLATE_FAST_BITS

INFLATE_FAST_LEN_SHIFT

INFLATE_FAST_VAL_MASK

INFLATE_MAX_BITS

INFLATE_MAX_CODELENS

INFLATE_MAX_DISTS

INFLATE_MAX_INPUT

INFLATE_MAX_LITS

INFLATE_MAX_LITS_DISTS

INFLATE_WINDOW_MASK

INFLATE_WINDOW_SIZE

INPUTWIDGET_LEN

INPUTWIDGET_MAX_LINES

INPUT_DEVICE_GAMEPAD

INPUT_DEVICE_NORMAL

INPUT_DEVICE_TOUCH

INPUT_MAX_GAMEPADS

INPUT_MAX_POINTERS

INPUT_SOURCE_GAMEPAD

INPUT_SOURCE_NORMAL

INVENTORY_BLOCKS_PER_HOTBAR

INVENTORY_HOTBARS

ISOMETRICDRAWER_MAXVERTICES

InputBind__BIND_AUTOROTATE

InputBind__BIND_AXIS_LINES

InputBind__BIND_BACK

InputBind__BIND_BREAK_LIQUIDS

InputBind__BIND_CHAT

InputBind__BIND_COUNT

InputBind__BIND_DELETE_BLOCK

InputBind__BIND_DROP_BLOCK

InputBind__BIND_EXT_INPUT

InputBind__BIND_FLY

InputBind__BIND_FLY_DOWN

InputBind__BIND_FLY_UP

InputBind__BIND_FOG

InputBind__BIND_FORWARD

InputBind__BIND_FULLSCREEN

InputBind__BIND_HALF_SPEED

InputBind__BIND_HIDE_FPS

InputBind__BIND_HIDE_GUI

InputBind__BIND_HOTBAR_1

InputBind__BIND_HOTBAR_2

InputBind__BIND_HOTBAR_3

InputBind__BIND_HOTBAR_4

InputBind__BIND_HOTBAR_5

InputBind__BIND_HOTBAR_6

InputBind__BIND_HOTBAR_7

InputBind__BIND_HOTBAR_8

InputBind__BIND_HOTBAR_9

InputBind__BIND_HOTBAR_LEFT

InputBind__BIND_HOTBAR_RIGHT

InputBind__BIND_HOTBAR_SWITCH

InputBind__BIND_IDOVERLAY

InputBind__BIND_INVENTORY

InputBind__BIND_JUMP

InputBind__BIND_LEFT

InputBind__BIND_LOOK_DOWN

InputBind__BIND_LOOK_LEFT

InputBind__BIND_LOOK_RIGHT

InputBind__BIND_LOOK_UP

InputBind__BIND_NOCLIP

InputBind__BIND_PICK_BLOCK

InputBind__BIND_PLACE_BLOCK

InputBind__BIND_RESPAWN

InputBind__BIND_RIGHT

InputBind__BIND_SCREENSHOT

InputBind__BIND_SEND_CHAT

InputBind__BIND_SET_SPAWN

InputBind__BIND_SMOOTH_CAMERA

InputBind__BIND_SPEED

InputBind__BIND_TABLIST

InputBind__BIND_THIRD_PERSON

InputBind__BIND_ZOOM_SCROLL

InputButtons_CCKEY_0

InputButtons_CCKEY_1

InputButtons_CCKEY_2

InputButtons_CCKEY_3

InputButtons_CCKEY_4

InputButtons_CCKEY_5

InputButtons_CCKEY_6

InputButtons_CCKEY_7

InputButtons_CCKEY_8

InputButtons_CCKEY_9

InputButtons_CCKEY_A

InputButtons_CCKEY_B

InputButtons_CCKEY_BACKSLASH

InputButtons_CCKEY_BACKSPACE

InputButtons_CCKEY_BROWSER_FAVORITES

InputButtons_CCKEY_BROWSER_HOME

InputButtons_CCKEY_BROWSER_NEXT

InputButtons_CCKEY_BROWSER_PREV

InputButtons_CCKEY_BROWSER_REFRESH

InputButtons_CCKEY_BROWSER_SEARCH

InputButtons_CCKEY_BROWSER_STOP

InputButtons_CCKEY_C

InputButtons_CCKEY_CAPSLOCK

InputButtons_CCKEY_COMMA

InputButtons_CCKEY_D

InputButtons_CCKEY_DELETE

InputButtons_CCKEY_DOWN

InputButtons_CCKEY_E

InputButtons_CCKEY_END

InputButtons_CCKEY_ENTER

InputButtons_CCKEY_EQUALS

InputButtons_CCKEY_ESCAPE

InputButtons_CCKEY_F

InputButtons_CCKEY_F1

InputButtons_CCKEY_F2

InputButtons_CCKEY_F3

InputButtons_CCKEY_F4

InputButtons_CCKEY_F5

InputButtons_CCKEY_F6

InputButtons_CCKEY_F7

InputButtons_CCKEY_F8

InputButtons_CCKEY_F9

InputButtons_CCKEY_F10

InputButtons_CCKEY_F11

InputButtons_CCKEY_F12

InputButtons_CCKEY_F13

InputButtons_CCKEY_F14

InputButtons_CCKEY_F15

InputButtons_CCKEY_F16

InputButtons_CCKEY_F17

InputButtons_CCKEY_F18

InputButtons_CCKEY_F19

InputButtons_CCKEY_F20

InputButtons_CCKEY_F21

InputButtons_CCKEY_F22

InputButtons_CCKEY_F23

InputButtons_CCKEY_F24

InputButtons_CCKEY_G

InputButtons_CCKEY_H

InputButtons_CCKEY_HOME

InputButtons_CCKEY_I

InputButtons_CCKEY_INSERT

InputButtons_CCKEY_J

InputButtons_CCKEY_K

InputButtons_CCKEY_KP0

InputButtons_CCKEY_KP1

InputButtons_CCKEY_KP2

InputButtons_CCKEY_KP3

InputButtons_CCKEY_KP4

InputButtons_CCKEY_KP5

InputButtons_CCKEY_KP6

InputButtons_CCKEY_KP7

InputButtons_CCKEY_KP8

InputButtons_CCKEY_KP9

InputButtons_CCKEY_KP_DECIMAL

InputButtons_CCKEY_KP_DIVIDE

InputButtons_CCKEY_KP_ENTER

InputButtons_CCKEY_KP_MINUS

InputButtons_CCKEY_KP_MULTIPLY

InputButtons_CCKEY_KP_PLUS

InputButtons_CCKEY_L

InputButtons_CCKEY_LALT

InputButtons_CCKEY_LAUNCH_APP1

InputButtons_CCKEY_LAUNCH_CALC

InputButtons_CCKEY_LAUNCH_MAIL

InputButtons_CCKEY_LAUNCH_MEDIA

InputButtons_CCKEY_LBRACKET

InputButtons_CCKEY_LCTRL

InputButtons_CCKEY_LEFT

InputButtons_CCKEY_LSHIFT

InputButtons_CCKEY_LWIN

InputButtons_CCKEY_M

InputButtons_CCKEY_MEDIA_NEXT

InputButtons_CCKEY_MEDIA_PLAY

InputButtons_CCKEY_MEDIA_PREV

InputButtons_CCKEY_MEDIA_STOP

InputButtons_CCKEY_MENU

InputButtons_CCKEY_MINUS

InputButtons_CCKEY_N

InputButtons_CCKEY_NUMLOCK

InputButtons_CCKEY_O

InputButtons_CCKEY_P

InputButtons_CCKEY_PAGEDOWN

InputButtons_CCKEY_PAGEUP

InputButtons_CCKEY_PAUSE

InputButtons_CCKEY_PERIOD

InputButtons_CCKEY_PRINTSCREEN

InputButtons_CCKEY_Q

InputButtons_CCKEY_QUOTE

InputButtons_CCKEY_R

InputButtons_CCKEY_RALT

InputButtons_CCKEY_RBRACKET

InputButtons_CCKEY_RCTRL

InputButtons_CCKEY_RIGHT

InputButtons_CCKEY_RSHIFT

InputButtons_CCKEY_RWIN

InputButtons_CCKEY_S

InputButtons_CCKEY_SCROLLLOCK

InputButtons_CCKEY_SEMICOLON

InputButtons_CCKEY_SLASH

InputButtons_CCKEY_SLEEP

InputButtons_CCKEY_SPACE

InputButtons_CCKEY_T

InputButtons_CCKEY_TAB

InputButtons_CCKEY_TILDE

InputButtons_CCKEY_U

InputButtons_CCKEY_UP

InputButtons_CCKEY_V

InputButtons_CCKEY_VOLUME_DOWN

InputButtons_CCKEY_VOLUME_MUTE

InputButtons_CCKEY_VOLUME_UP

InputButtons_CCKEY_W

InputButtons_CCKEY_X

InputButtons_CCKEY_Y

InputButtons_CCKEY_Z

InputButtons_CCMOUSE_L

InputButtons_CCMOUSE_M

InputButtons_CCMOUSE_R

InputButtons_CCMOUSE_X1

InputButtons_CCMOUSE_X2

InputButtons_CCMOUSE_X3

InputButtons_CCMOUSE_X4

InputButtons_CCMOUSE_X5

InputButtons_CCMOUSE_X6

InputButtons_CCPAD_1

InputButtons_CCPAD_2

InputButtons_CCPAD_3

InputButtons_CCPAD_4

InputButtons_CCPAD_5

InputButtons_CCPAD_6

InputButtons_CCPAD_7

InputButtons_CCPAD_CDOWN

InputButtons_CCPAD_CLEFT

InputButtons_CCPAD_CRIGHT

InputButtons_CCPAD_CUP

InputButtons_CCPAD_DOWN

InputButtons_CCPAD_L

InputButtons_CCPAD_LEFT

InputButtons_CCPAD_LSTICK

InputButtons_CCPAD_R

InputButtons_CCPAD_RIGHT

InputButtons_CCPAD_RSTICK

InputButtons_CCPAD_SELECT

InputButtons_CCPAD_START

InputButtons_CCPAD_UP

InputButtons_CCPAD_ZL

InputButtons_CCPAD_ZR

InputButtons_CCWHEEL_DOWN

InputButtons_CCWHEEL_LEFT

InputButtons_CCWHEEL_RIGHT

InputButtons_CCWHEEL_UP

InputButtons_INPUT_CLIPBOARD_COPY

InputButtons_INPUT_CLIPBOARD_PASTE

InputButtons_INPUT_COUNT

InputButtons_INPUT_NONE

Input_DisplayNames

Input_StorageNames

Input_TouchMode

Int16_MaxValue

Int32_MaxValue

Int32_MinValue

KB_B_BACK

KB_B_CAPS

KB_B_CLOSE

KB_B_ENTER

KB_B_SHIFT

KB_B_SPACE

KB_TILE_SIZE

KEYBOARD_FLAG_SEND

KeyBind_Defaults

Key_Ascii_Names

Key_Function_Names

KeyboardType_KEYBOARD_TYPE_INTEGER

KeyboardType_KEYBOARD_TYPE_NUMBER

KeyboardType_KEYBOARD_TYPE_PASSWORD

KeyboardType_KEYBOARD_TYPE_TEXT

LIGHT_FLAG_SHADES_FROM_BELOW

LINPUT_HEIGHT

LLAYOUT_EXTRA

LLAYOUT_HEIGHT

LLAYOUT_WIDTH

LLINE_HEIGHT

LOPT_AUTO_CLOSE

LOPT_PASSWORD

LOPT_SESSION

LOPT_SHOW_EMPTY

LOPT_USERNAME

LU_HAS_PITCH

LU_HAS_POS

LU_HAS_ROTX

LU_HAS_ROTZ

LU_HAS_YAW

LU_ORI_INTERPOLATE

LU_POS_ABSOLUTE_INSTANT

LU_POS_ABSOLUTE_SMOOTH

LU_POS_MODEMASK

LU_POS_RELATIVE_SHIFT

LU_POS_RELATIVE_SMOOTH

LWIDGET_TYPE_LWIDGET_BUTTON

LWIDGET_TYPE_LWIDGET_CHECKBOX

LWIDGET_TYPE_LWIDGET_INPUT

LWIDGET_TYPE_LWIDGET_LABEL

LWIDGET_TYPE_LWIDGET_LINE

LWIDGET_TYPE_LWIDGET_SLIDER

LWIDGET_TYPE_LWIDGET_TABLE

LightingMode_LIGHTING_MODE_CLASSIC

LightingMode_LIGHTING_MODE_COUNT

LightingMode_LIGHTING_MODE_FANCY

MATH_DEG2RAD

MATH_LARGENUM

MATH_PI

MATH_RAD2DEG

MAX_CUSTOM_MODELS

MAX_CUSTOM_MODEL_ANIMS

MAX_CUSTOM_MODEL_PARTS

MAX_LOCAL_PLAYERS

MAX_NET_PLAYERS

MENUINPUTWIDGET_MAX

MILLIS_PER_SEC

MODEL_BOX_VERTICES

MODEL_FLAG_CLEAR_HAT

MODEL_FLAG_INITED

MODEL_QUAD_VERTICES

MatrixType__MATRIX_PROJ

MatrixType__MATRIX_VIEW

Matrix_Identity

Identity matrix. (A * Identity = A)

MsgType_MSG_TYPE_ANNOUNCEMENT

MsgType_MSG_TYPE_BIGANNOUNCEMENT

MsgType_MSG_TYPE_BOTTOMRIGHT_1

MsgType_MSG_TYPE_BOTTOMRIGHT_2

MsgType_MSG_TYPE_BOTTOMRIGHT_3

MsgType_MSG_TYPE_CLIENTSTATUS_1

MsgType_MSG_TYPE_CLIENTSTATUS_2

MsgType_MSG_TYPE_EXTRASTATUS_1

MsgType_MSG_TYPE_EXTRASTATUS_2

MsgType_MSG_TYPE_NORMAL

MsgType_MSG_TYPE_SMALLANNOUNCEMENT

MsgType_MSG_TYPE_STATUS_1

MsgType_MSG_TYPE_STATUS_2

MsgType_MSG_TYPE_STATUS_3

NATIVE_STR_LEN

NameMode_NAME_MODE_ALL

NameMode_NAME_MODE_ALL_HOVERED

NameMode_NAME_MODE_ALL_UNSCALED

NameMode_NAME_MODE_COUNT

NameMode_NAME_MODE_HOVERED

NameMode_NAME_MODE_NONE

OFD_UPLOAD_DELETE

OFD_UPLOAD_PERSIST

OGG_BUFFER_SIZE

OPCODE__OPCODE_ADD_ENTITY

OPCODE__OPCODE_BULK_BLOCK_UPDATE

OPCODE__OPCODE_CINEMATIC_GUI

OPCODE__OPCODE_COUNT

OPCODE__OPCODE_CUSTOM_BLOCK_LEVEL

OPCODE__OPCODE_DEFINE_BLOCK

OPCODE__OPCODE_DEFINE_BLOCK_EXT

OPCODE__OPCODE_DEFINE_EFFECT

OPCODE__OPCODE_DEFINE_MODEL

OPCODE__OPCODE_DEFINE_MODEL_PART

OPCODE__OPCODE_ENTITY_TELEPORT

OPCODE__OPCODE_ENTITY_TELEPORT_EXT

OPCODE__OPCODE_ENV_SET_COLOR

OPCODE__OPCODE_ENV_SET_MAP_APPEARANCE

OPCODE__OPCODE_ENV_SET_MAP_PROPERTY

OPCODE__OPCODE_ENV_SET_MAP_URL

OPCODE__OPCODE_ENV_SET_WEATHER

OPCODE__OPCODE_EXT_ADD_ENTITY

OPCODE__OPCODE_EXT_ADD_ENTITY2

OPCODE__OPCODE_EXT_ADD_PLAYER_NAME

OPCODE__OPCODE_EXT_ENTRY

OPCODE__OPCODE_EXT_INFO

OPCODE__OPCODE_EXT_REMOVE_PLAYER_NAME

OPCODE__OPCODE_HACK_CONTROL

OPCODE__OPCODE_HANDSHAKE

OPCODE__OPCODE_HOLD_THIS

OPCODE__OPCODE_KICK

OPCODE__OPCODE_LEVEL_BEGIN

OPCODE__OPCODE_LEVEL_DATA

OPCODE__OPCODE_LEVEL_END

OPCODE__OPCODE_LIGHTING_MODE

OPCODE__OPCODE_MAKE_SELECTION

OPCODE__OPCODE_MESSAGE

OPCODE__OPCODE_ORI_UPDATE

OPCODE__OPCODE_PING

OPCODE__OPCODE_PLAYER_CLICK

OPCODE__OPCODE_PLUGIN_MESSAGE

OPCODE__OPCODE_RELPOS_AND_ORI_UPDATE

OPCODE__OPCODE_RELPOS_UPDATE

OPCODE__OPCODE_REMOVE_ENTITY

OPCODE__OPCODE_REMOVE_SELECTION

OPCODE__OPCODE_SET_BLOCK

OPCODE__OPCODE_SET_BLOCK_CLIENT

OPCODE__OPCODE_SET_BLOCK_PERMISSION

OPCODE__OPCODE_SET_ENTITY_PROPERTY

OPCODE__OPCODE_SET_HOTBAR

OPCODE__OPCODE_SET_INVENTORY_ORDER

OPCODE__OPCODE_SET_MODEL

OPCODE__OPCODE_SET_PERMISSION

OPCODE__OPCODE_SET_REACH

OPCODE__OPCODE_SET_SPAWNPOINT

OPCODE__OPCODE_SET_TEXT_COLOR

OPCODE__OPCODE_SET_TEXT_HOTKEY

OPCODE__OPCODE_SPAWN_EFFECT

OPCODE__OPCODE_TWO_WAY_PING

OPCODE__OPCODE_UNDEFINE_BLOCK

OPCODE__OPCODE_UNDEFINE_MODEL

OPCODE__OPCODE_VELOCITY_CONTROL

OPT_ANAGLYPH3D

OPT_BLACK_TEXT

OPT_BLOCK_PHYSICS

OPT_CAMERA_CLIPPING

OPT_CAMERA_MASS

OPT_CAMERA_SMOOTH

OPT_CHATLINES

OPT_CHAT_AUTO_SCALE

OPT_CHAT_LOGGING

OPT_CHAT_SCALE

OPT_CLASSIC_ARM_MODEL

OPT_CLASSIC_CHAT

OPT_CLASSIC_GUI

OPT_CLASSIC_HACKS

OPT_CLASSIC_INVENTORY

OPT_CLASSIC_MODE

OPT_CLASSIC_OPTIONS

OPT_CLASSIC_TABLIST

OPT_CLICKABLE_CHAT

OPT_CPE

OPT_CROSSHAIR_SCALE

OPT_CUSTOM_BLOCKS

OPT_DEFAULT_TEX_PACK

OPT_DPI_SCALING

OPT_ENTITY_SHADOW

OPT_FIELD_OF_VIEW

OPT_FONT_NAME

OPT_FORCE_OPENAL

OPT_FPS_LIMIT

OPT_FULL_BLOCK_STEP

OPT_GAME_VERSION

OPT_GRAB_CURSOR

OPT_HACKS_ENABLED

OPT_HACK_PERM_MSGS

OPT_HOTBAR_SCALE

OPT_HTTPS_VERIFY

OPT_HTTP_ONLY

OPT_INVENTORY_SCALE

OPT_INVERT_MOUSE

OPT_INV_SCROLLBAR_SCALE

OPT_JUMP_VELOCITY

OPT_LANDSCAPE_MODE

OPT_LIGHTING_MODE

OPT_MAX_CHUNK_UPDATES

OPT_MAX_MUSIC_DELAY

OPT_MIN_MUSIC_DELAY

OPT_MIPMAPS

OPT_MODIFIABLE_LIQUIDS

OPT_MUSIC_VOLUME

OPT_NAMES_MODE

OPT_NOCLIP_SLIDE

OPT_PUSHBACK_PLACING

OPT_RAW_INPUT

OPT_RENDER_TYPE

OPT_SELECTED_BLOCK_OUTLINE_COLOR

OPT_SELECTED_BLOCK_OUTLINE_OPACITY

OPT_SELECTED_BLOCK_OUTLINE_SCALE

OPT_SENSITIVITY

OPT_SERVER_TEXTURES

OPT_SHOW_BLOCK_IN_HAND

OPT_SHOW_FPS

OPT_SIMPLE_ARMS_ANIM

OPT_SKIN_SERVER

OPT_SMOOTH_LIGHTING

OPT_SOUND_VOLUME

OPT_SPEED_FACTOR

OPT_TAB_AUTOCOMPLETE

OPT_TOUCH_BUTTONS

OPT_TOUCH_HALIGN

OPT_TOUCH_SCALE

OPT_USE_CHAT_FONT

OPT_VIEW_BOBBING

OPT_VIEW_DISTANCE

OPT_WINDOW_HEIGHT

OPT_WINDOW_WIDTH

OPT_WOM_STYLE_HACKS

PACKEDCOL_A_MASK

PACKEDCOL_A_SHIFT

PACKEDCOL_B_MASK

PACKEDCOL_B_SHIFT

PACKEDCOL_G_MASK

PACKEDCOL_G_SHIFT

PACKEDCOL_RGB_MASK

PACKEDCOL_R_MASK

PACKEDCOL_R_SHIFT

PACKEDCOL_SHADE_X

PACKEDCOL_SHADE_YMIN

PACKEDCOL_SHADE_Z

PACKEDCOL_WHITE

PAD_AXIS_PAD_AXIS_LEFT

PAD_AXIS_PAD_AXIS_RIGHT

PNG_MAX_DIMS

PNG_SIG_SIZE

PROTOCOL_VERSION__PROTOCOL_0017

PROTOCOL_VERSION__PROTOCOL_0019

PROTOCOL_VERSION__PROTOCOL_0020

PROTOCOL_VERSION__PROTOCOL_0030

PadBind_Defaults

Pad_DisplayNames

Pad_Names

Pointers_Count

REGISTERNEW_URL

RESOURCE_SERVER

RESPAWN_NOT_FOUND

ROPT_IP

ROPT_MPPASS

ROPT_PORT

ROPT_SERVER

ROPT_USER

RotateOrder_ROTATE_ORDER_XYZ

RotateOrder_ROTATE_ORDER_XZY

RotateOrder_ROTATE_ORDER_YZX

RotateOrder_ROTATE_ORDER_ZYX

SECS_PER_DAY

SECS_PER_HOUR

SECS_PER_MIN

SERVICES_SERVER

SIZEOF_VERTEX_COLOURED

SIZEOF_VERTEX_TEXTURED

SKINS_SERVER

SKIN_FETCH_COMPLETED

SKIN_FETCH_DOWNLOADING

SKIN_TYPE_SKIN_64x32

SKIN_TYPE_SKIN_64x64

SKIN_TYPE_SKIN_64x64_SLIM

SKIN_TYPE_SKIN_INVALID

SOCKET_MAX_ADDRS

SOPT_SERVICES

STRINGSBUFFER_BUFFER_DEF_SIZE

STRINGSBUFFER_DEF_LEN_MASK

STRINGSBUFFER_DEF_LEN_SHIFT

STRINGSBUFFER_FLAGS_DEF_ELEMS

STRING_INT_CHARS

STRING_SIZE

Screen3DS_BOTTOM_SCREEN

Screen3DS_TOP_SCREEN

ShadowMode_SHADOW_MODE_CIRCLE

ShadowMode_SHADOW_MODE_CIRCLE_ALL

ShadowMode_SHADOW_MODE_COUNT

ShadowMode_SHADOW_MODE_NONE

ShadowMode_SHADOW_MODE_SNAP_TO_BLOCK

Socket_PollMode_SOCKET_POLL_READ

Socket_PollMode_SOCKET_POLL_WRITE

SoftKeyboard_SOFT_KEYBOARD_NONE

SoftKeyboard_SOFT_KEYBOARD_RESIZE

SoftKeyboard_SOFT_KEYBOARD_SHIFT

SoftKeyboard_SOFT_KEYBOARD_VIRTUAL

SoundType_SOUND_CLOTH

SoundType_SOUND_COUNT

SoundType_SOUND_GLASS

SoundType_SOUND_GRASS

SoundType_SOUND_GRAVEL

SoundType_SOUND_METAL

SoundType_SOUND_NONE

SoundType_SOUND_SAND

SoundType_SOUND_SNOW

SoundType_SOUND_STONE

SoundType_SOUND_WOOD

TABLE_MAX_VERTICES

TABLIST_MAX_NAMES

TEXTATLAS_MAX_WIDTHS

TEXTGROUPWIDGET_LEN

TEXTURE_FLAG_BILINEAR

TEXTURE_FLAG_DYNAMIC

TEXTURE_FLAG_LOWRES

TEXTURE_FLAG_MANAGED

TEXTURE_FLAG_NONPOW2

TEXTWIDGET_MAX

TOUCH_TYPE_ALL

TOUCH_TYPE_BLOCKS

TOUCH_TYPE_CAMERA

TOUCH_TYPE_GUI

TREE_MAX_COUNT

UInt8_MaxValue

UInt16_MaxValue

UNIX_EPOCH_SECONDS

UPDATES_SERVER

UPDATE_FILE

URL_MAX_SIZE

UV2_Scale

VORBIS_MAX_BLOCK_SIZE

VORBIS_MAX_CHANS

VertexFormat__VERTEX_FORMAT_COLOURED

VertexFormat__VERTEX_FORMAT_TEXTURED

WIDGET_FLAG_DISABLED

WIDGET_FLAG_MAINSCREEN

WIDGET_FLAG_SELECTABLE

WORLD_UUID_LEN

Weather__WEATHER_RAINY

Weather__WEATHER_SNOWY

Weather__WEATHER_SUNNY

WindowState_WINDOW_STATE_FULLSCREEN

WindowState_WINDOW_STATE_MINIMISED

WindowState_WINDOW_STATE_NORMAL

_NL

bindNames

controlChars

extendedChars

Statics

Atlas1D⚠

Atlas2D⚠

BlockEvents⚠

Blocks⚠

Camera⚠

ChatEvents⚠

ControllerEvents⚠

DisplayInfo⚠

Drawer⚠

Drawer2D⚠

Entities⚠

EntityEvents⚠

Env⚠

EventAPIVersion⚠

Game⚠

Gfx⚠

GfxEvents⚠

Gui⚠

InputEvents⚠

Inventory⚠

Lighting⚠

Models⚠

NetEvents⚠

Physics⚠

PointerEvents⚠

Pointers⚠

Protocol⚠

Server⚠

TabList⚠

TabListEvents⚠

TextureEvents⚠

UserEvents⚠

WindowEvents⚠

WindowInfo⚠

World⚠

WorldEvents⚠

Functions

AABB_Make⚠

Bitmap_Scale⚠

Block_FindID⚠

Block_Parse⚠

Block_UNSAFE_GetName⚠

BoxDesc_BuildBox⚠

BoxDesc_BuildRotatedBox⚠

BoxDesc_XQuad⚠

BoxDesc_XQuad2⚠

BoxDesc_YQuad⚠

BoxDesc_YQuad2⚠

BoxDesc_ZQuad⚠

BoxDesc_ZQuad2⚠

CPE_SendPluginMessage⚠

Camera_Register⚠

Chat_Add⚠

Chat_Add1⚠

Chat_Add2⚠

Chat_Add3⚠

Chat_Add4⚠

Chat_AddOf⚠

Chat_AddRaw⚠

Chat_Send⚠

Clipboard_GetText⚠

Clipboard_SetText⚠

Commands_Register⚠

Context2D_Alloc⚠

Context2D_Clear⚠

Context2D_DrawPixels⚠

Context2D_DrawText⚠

Context2D_Free⚠

Context2D_MakeTexture⚠

Context2D_Wrap⚠

Convert_CP437ToUnicode

Convert_CodepointToCP437

Convert_ParseBool⚠

Convert_ParseFloat⚠

Convert_ParseInt⚠

Convert_ParseUInt8⚠

Convert_ParseUInt16⚠

Convert_ParseUInt64⚠

Convert_TryCodepointToCP437⚠

DateTime_CurrentLocal⚠

DateTime_CurrentUTC⚠

Deflate_MakeStream⚠

Directory_Enum⚠

Drawer2D_MakeTextTexture⚠

Drawer2D_TextHeight⚠

Drawer2D_TextWidth⚠

Drawer_XMax⚠

Drawer_XMin⚠

Drawer_YMax⚠

Drawer_YMin⚠

Drawer_ZMax⚠

Drawer_ZMin⚠

DynamicLib_DescribeError⚠

DynamicLib_Get⚠

DynamicLib_Get2⚠

DynamicLib_Load⚠

DynamicLib_Load2⚠

Entity_GetTransform⚠

Entity_Init⚠

Entity_SetModel⚠

Entity_TouchesAny⚠

Env_Reset⚠

Env_SetCloudsCol⚠

Env_SetCloudsHeight⚠

Env_SetCloudsSpeed⚠

Env_SetEdgeBlock⚠

Env_SetEdgeHeight⚠

Env_SetExpFog⚠

Env_SetFogCol⚠

Env_SetLampLightCol⚠

Env_SetLavaLightCol⚠

Env_SetShadowCol⚠

Env_SetSidesBlock⚠

Env_SetSidesOffset⚠

Env_SetSkyCol⚠

Env_SetSkyboxCol⚠

Env_SetSkyboxHorSpeed⚠

Env_SetSkyboxVerSpeed⚠

Env_SetSunCol⚠

Env_SetWeather⚠

Env_SetWeatherFade⚠

Env_SetWeatherSpeed⚠

Event_RaiseBlock⚠

Calls all registered callbacks for an event which takes block change arguments. These are the coordinates/location of the change, block there before, block there now.

Event_RaiseChat⚠

Calls all registered callbacks for an event which has chat message type and contents. See MsgType enum in Chat.h for what types of messages there are.

Event_RaiseEntry⚠

Calls all registered callbacks for an event which has data stream and name arguments.

Event_RaiseFloat⚠

Event_RaiseInput⚠

Calls all registered callbacks for an event which has keyboard key/mouse button. repeating is whether the key/button was already pressed down. (i.e. user is holding down key)

Event_RaiseInt⚠

Event_RaiseLightingMode⚠

Calls all registered callbacks for an event called when the Lighting_LightingMode is changed

Event_RaisePadAxis⚠

Calls all registered callbacks for an event which has pad axis arguments.

Event_RaisePluginMessage⚠

Calls all registered callbacks for an event which has a channel and a 64 byte data argument.

Event_RaiseRawMove⚠

Calls all registered callbacks for an event which has raw pointer movement arguments.

Event_RaiseString⚠

Calls all registered callbacks for an event which has a string argument.

Event_RaiseVoid⚠

Event_Register⚠

Event_RegisterBlock⚠

Event_RegisterChat⚠

Event_RegisterEntry⚠

Event_RegisterFloat⚠

Event_RegisterInput⚠

Event_RegisterInt⚠

Event_RegisterLightingMode⚠

Event_RegisterPadAxis⚠

Event_RegisterPluginMessage⚠

Event_RegisterRawMove⚠

Event_RegisterString⚠

Event_RegisterVoid⚠

Event_Unregister⚠

Event_UnregisterBlock⚠

Event_UnregisterChat⚠

Event_UnregisterEntry⚠

Event_UnregisterFloat⚠

Event_UnregisterInput⚠

Event_UnregisterInt⚠

Event_UnregisterLightingMode⚠

Event_UnregisterPadAxis⚠

Event_UnregisterPluginMessage⚠

Event_UnregisterRawMove⚠

Event_UnregisterString⚠

Event_UnregisterVoid⚠

Font_Free⚠

Font_Make⚠

GZip_MakeStream⚠

Game_ChangeBlock⚠

Game_UpdateBlock⚠

Gfx_BindDynamicVb⚠

Gfx_BindIb⚠

Gfx_BindTexture⚠

Gfx_BindVb⚠

Gfx_ClearBuffers⚠

Gfx_ClearColor⚠

Gfx_CreateDynamicVb⚠

Gfx_CreateIb2⚠

Gfx_CreateTexture⚠

Gfx_CreateVb⚠

Gfx_DeleteDynamicVb⚠

Gfx_DeleteIb⚠

Gfx_DeleteTexture⚠

Gfx_DeleteVb⚠

Gfx_DepthOnlyRendering⚠

Gfx_DisableMipmaps⚠

Gfx_DisableTextureOffset⚠

Gfx_DrawVb_IndexedTris⚠

Gfx_DrawVb_IndexedTris_Range⚠

Gfx_DrawVb_Lines⚠

Gfx_EnableMipmaps⚠

Gfx_EnableTextureOffset⚠

Gfx_GetFog⚠

Gfx_LoadMatrix⚠

Gfx_LockDynamicVb⚠

Gfx_LockVb⚠

Gfx_Make2DQuad

Gfx_SetAlphaArgBlend⚠

Gfx_SetAlphaBlending⚠

Gfx_SetAlphaTest⚠

Gfx_SetColorWrite⚠

Gfx_SetDepthTest⚠

Gfx_SetDepthWrite⚠

Gfx_SetDynamicVbData⚠

Gfx_SetFaceCulling⚠

Gfx_SetFog⚠

Gfx_SetFogCol⚠

Gfx_SetFogDensity⚠

Gfx_SetFogEnd⚠

Gfx_SetFogMode⚠

Gfx_SetScissor⚠

Gfx_SetTexturing⚠

Gfx_SetVertexFormat⚠

Gfx_SetViewport⚠

Gfx_UnlockDynamicVb⚠

Gfx_UnlockVb⚠

Gfx_UpdateDynamicVb_IndexedTris⚠

Gfx_UpdateTexturePart⚠

Gradient_Blend⚠

Gradient_Noise⚠

Gradient_Vertical⚠

Gui_Add⚠

Gui_GetInputGrab⚠

Gui_GetScreen⚠

Gui_Remove⚠

IVec3_Floor

IVec3_Max

IVec3_MaxValue

Returns a vector with all components set to Int32_MaxValue.

IVec3_Min

IVec3_ToVec3

Inflate_Init2⚠

Inflate_MakeStream2⚠

Inventory_Get

Gets the block at the nth index in the current hotbar.

Inventory_SelectedBlock

Gets the currently selected block.

Inventory_Set

Sets the block at the nth index in the current hotbar.

KeyBind_IsPressed⚠

MapImporter_Find⚠

MapImporter_Register⚠

Map_LoadFrom⚠

Math_Cos⚠

Math_CosF⚠

Math_IsPowOf2

Math_NextPowOf2

Math_Sin⚠

Math_SinF⚠

Matrix_IdentityValue

Matrix_LookRot

Matrix_Mul⚠

Matrix_Orthographic

Matrix_PerspectiveFieldOfView

Matrix_RotateX⚠

Matrix_RotateY⚠

Matrix_RotateZ⚠

Matrix_Scale⚠

Matrix_Translate⚠

Mem_Alloc⚠

Mem_AllocCleared⚠

Mem_Free⚠

Mem_Realloc⚠

Mem_TryAlloc⚠

Mem_TryAllocCleared⚠

Mem_TryRealloc⚠

ModelPart_Init

Model_ApplyTexture⚠

Model_DrawArmPart⚠

Model_DrawPart⚠

Model_DrawRotate⚠

Model_Get⚠

Model_Init⚠

Model_Register⚠

Model_RegisterTexture⚠

Model_Render⚠

Model_SetupState⚠

Model_UpdateVB⚠

Mutex_Create⚠

Mutex_Free⚠

Mutex_Lock⚠

Mutex_Unlock⚠

NetPlayer_Init⚠

Options_Get⚠

Options_GetBool⚠

Options_GetEnum⚠

Options_GetFloat⚠

Options_GetInt⚠

Options_Reload⚠

Options_SaveIfChanged⚠

Options_Set⚠

Options_SetBool⚠

Options_SetInt⚠

Options_SetString⚠

PackedCol_A

PackedCol_A_Bits

PackedCol_B

PackedCol_B_Bits

PackedCol_G

PackedCol_G_Bits

PackedCol_Lerp⚠

PackedCol_Make

PackedCol_R

PackedCol_R_Bits

PackedCol_Scale⚠

PackedCol_ScreenBlend⚠

PackedCol_Tint⚠

Particle_DoRender

Png_Decode⚠

Process_Exit⚠

Process_StartGame2⚠

Process_StartOpen⚠

Random_Float⚠

Random_Next⚠

Random_Seed⚠

ScheduledTask_Add⚠

Selections_Add⚠

Selections_Remove⚠

Stopwatch_ElapsedMicroseconds⚠

Stopwatch_Measure⚠

Stream_CreateFile⚠

Stream_FromFile⚠

Stream_OpenFile⚠

Stream_Read⚠

Stream_ReadLine⚠

Stream_ReadonlyBuffered⚠

Stream_ReadonlyMemory⚠

Stream_ReadonlyPortion⚠

Stream_Write⚠

Stream_WriteLine⚠

String_Append⚠

String_AppendBool⚠

String_AppendCP1252⚠

String_AppendColorless⚠

String_AppendConst⚠

String_AppendFloat⚠

String_AppendHex⚠

String_AppendInt⚠

String_AppendPaddedInt⚠

String_AppendString⚠

String_AppendUInt32⚠

String_AppendUtf8⚠

String_AppendUtf16⚠

String_CalcLen⚠

String_CaselessContains⚠

String_CaselessEnds⚠

String_CaselessEquals⚠

String_CaselessEqualsConst⚠

String_CaselessStarts⚠

String_Compare⚠

String_Copy⚠

String_CopyToRaw⚠

String_CopyToRawArray

Calls String_CopyToRaw on a compile time array.

String_DeleteAt⚠

String_EncodeUtf8⚠

String_Equals⚠

String_Format1⚠

String_Format2⚠

String_Format3⚠

String_Format4⚠

String_FromRaw⚠

Constructs a string from a (maybe null terminated) buffer.

String_FromRawArray⚠

Constructs a string from a compile time array, that may have arbitary actual length of data at runtime

String_FromReadonly⚠

String_IndexOfAt⚠

String_IndexOfConst⚠

String_Init⚠

Constructs a string from the given arguments.

String_InsertAt⚠

String_LastIndexOfAt⚠

String_UNSAFE_Separate⚠

String_UNSAFE_Split⚠

String_UNSAFE_SplitBy⚠

String_UNSAFE_Substring⚠

String_UNSAFE_SubstringAt⚠

String_UNSAFE_TrimEnd⚠

String_UNSAFE_TrimStart⚠

StringsBuffer_Add⚠

StringsBuffer_Remove⚠

StringsBuffer_UNSAFE_Get⚠

SysFonts_GetNames⚠

TabList_Remove⚠

TabList_Set⚠

Tan_Simple

TexturePack_Extract⚠

Thread_Detach⚠

Thread_Join⚠

Thread_Run⚠

Thread_Sleep⚠

UNSAFE_GetString⚠

Vec3_BigPos

Vec3_Create3

Vec3_Equals

Whether all of the components of the two vectors are equal.

Vec3_GetDirVector

Returns a normalised vector facing in the direction described by the given yaw and pitch.

Vec3_LengthSquared

Returns the squared length of the vector. Squared length can be used for comparison, to avoid a costly sqrt() However, you must sqrt() this when adding lengths.

Vec3_Lerp

Linearly interpolates components of two vectors.

Vec3_Normalize

Scales all components of a vector to lie in [-1, 1]

Vec3_RotateX

Vec3_RotateY

Vec3_RotateY3

Vec3_RotateZ

Vec3_Transform

Transforms a vector by the given matrix.

Vec3_TransformY

Same as Vec3_Transform, but faster since x and z are assumed as 0.

Waitable_Create⚠

Waitable_Free⚠

Waitable_Signal⚠

Waitable_Wait⚠

Waitable_WaitFor⚠

Window_SetTitle⚠

Window_ShowDialog⚠

World_Blocks

World_Blocks2

World_GetBlock

World_NewMap⚠

World_Pack

World_SetNewMap⚠

ZLib_MakeStream⚠

Type Aliases

AXIS_BEHAVIOUR

AXIS_SENSITIVITY

BLOCKID

BindMapping

BindReleased

BindTriggered

BitmapCol

BlockID

BlockRaw

Button_GetBool

Button_GetEnum

Button_GetHex

Button_GetInt

Button_GetNum

Button_SetBool

Button_SetEnum

Button_SetHex

Button_SetInt

Button_SetNum

CC_ERRORS

CollideType

CustomModelAnimAxis

CustomModelAnimType

DefaultZipCallback

Directory_EnumCallback

DrawType

EntityID

EntityType

Entity_TouchesCondition

EntryList_Filter

EnvVar

Event_Block_Callback

Event_Chat_Callback

Event_Entry_Callback

Event_Float_Callback

Event_Input_Callback

Event_Int_Callback

Event_LightingMode_Callback

Event_PadAxis_Callback

Event_PluginMessage_Callback

Event_RawMove_Callback

Event_String_Callback

Event_Void_Callback

FACE_CONSTS

FONT_FLAGS

FP_Block_UNSAFE_GetName

FP_Chat_Add

FP_Chat_AddOf

FP_Chat_Send

FP_Commands_Register

FP_GZip_MakeStream

FP_Random_Next

FP_Random_Seed

FP_Stream_CreateFile

FP_Stream_OpenFile

FP_Stream_Read

FP_Stream_Write

FP_String_AppendConst

FP_String_CaselessContains

FP_String_CaselessEnds

FP_String_CaselessEquals

FP_String_CaselessEqualsConst

FP_String_CaselessStarts

FP_String_Equals

FP_String_Format1

FP_String_Format2

FP_String_Format3

FP_String_Format4

FP_ZLib_MakeStream

Face

FallbackFont_Plotter

FetcherErrorCallback

FileDialogCallback

File_SeekFrom

FogFunc

FogFunc_

FpsLimitMethod

GAME_VERSION_

Game_Draw2DHook

GfxBuffers

GfxBuffers_

GfxResourceID

Gfx_FillIBFunc

GuiAnchor

GuiPriority

HotkeyModifiers

HttpProgress

HttpRequestType

IVec3

InputBind

InputBind_

InputButtons

InputDevice_IsPressed

JsonOnNew

JsonOnValue

KeyboardType

LBackend_DrawHook

LCheckboxChanged

LScreen_Func

LWIDGET_TYPE

LWebTask_ErrorCallback

LWidgetFunc

LightingMode

Logger_DescribeError

Logger_DoWarn

MapImportFunc

MatrixType

MatrixType_

MenuInputDone

MsgType

NameMode

Net_Handler

OPCODE_

PAD_AXIS

PROTOCOL_VERSION_

PackedCol

PhysicsHandler

Png_RowGetter

RNGState

Rect2D

RotateOrder

SKIN_TYPE

ScheduledTaskCallback

Screen3DS

ShadowMode

Socket_PollMode

SoftKeyboard

SoundType

SpecialInputAppendFunc

SysFont_RegisterCallback

TextGroupWidget_Get

TextureLoc

TextureRec

Thread_StartFunc

TimeMS

Vec2

Vec3

VertexFormat

VertexFormat_

Weather_

Widget_LeftClick

WindowState

Zip_ProcessEntry

Zip_SelectEntry

cc_bool

cc_codepoint

cc_file

cc_filepath

cc_int8

cc_int16

cc_int32

cc_int64

cc_pointer

cc_result

cc_sockaddr

cc_socket

cc_string

cc_uint8

cc_uint16

cc_uint32

cc_uint64

cc_uintptr

cc_unichar

cc_winstring

Unions

AudioChunkMeta

InflateState__bindgen_ty_1

IntAndFloat

MenuInputDesc__bindgen_ty_1

Stream__bindgen_ty_1

cc_pointer_